% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_smooth}
\alias{age_smooth}
\title{Smoothed age distributions by race and gender.}
\format{
A list of 10 elements named for unique combinations of NHANES gender
 and race/ethnicity categories. Each list element is a matrix with 401 rows and 2 columns:
 \describe{\item{\code{x}}{401 evenly-spaced grid
 points between 0 and 959 months} \item{\code{y}}{Smoothed density values
 evaluated at each \code{x} value} }
}
\usage{
age_smooth
}
\description{
Distributions of ages in months, computed from NHANES data smoothed using
\code{\link[survey]{svysmooth}}, for each combination of race/ethnicity and
gender.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
 environmental chemicals by simulating toxicokinetic variability."
 Environment International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{data}
\keyword{httk-pop}
