% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armitage.R
\name{armitage_eval}
\alias{armitage_eval}
\title{Armitage In Vitro Distribution Model}
\usage{
armitage_eval(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  casrn.vector = NA_character_,
  nomconc.vector = 1,
  this.well_number = 384,
  this.FBSf = NA_real_,
  tcdata = NA,
  user_assay_parameters = NA,
  this.sarea = NA_real_,
  this.v_total = NA_real_,
  this.v_working = NA_real_,
  this.cell_yield = NA_real_,
  this.Tsys = 37,
  this.Tref = 298.15,
  this.option.kbsa2 = FALSE,
  this.option.swat2 = FALSE,
  this.option.kpl2 = FALSE,
  this.option.bottom = TRUE,
  this.pseudooct = 0.01,
  this.memblip = 0.04,
  this.nlom = 0.05,
  this.P_nlom = 0.035,
  this.P_dom = 0.05,
  this.P_cells = 1,
  this.cell_pH = 7.4,
  this.Anionic_VF = 0.175,
  this.A_Prop_acid = 0.05,
  this.A_Prop_base = 20,
  this.Lyso_VF = 0.0068,
  this.Lyso_Diam = 500,
  this.Lyso_pH = 5.1,
  this.csalt = 0.15,
  this.celldensity = 1,
  this.cellmass = 3,
  this.f_oc = 1,
  this.conc_ser_alb = 24,
  this.conc_ser_lip = 1.9,
  this.Vdom = 0,
  this.pH = 7,
  restrict.ion.partitioning = FALSE,
  surface.area.switch = TRUE,
  chemdata = chem.physical_and_invitro.data
)
}
\arguments{
\item{chem.cas}{A single or vector of Chemical Abstracts Service Registry 
Number(s) (CAS-RN) of desired chemical(s).}

\item{chem.name}{A single or vector of name(s)) of desired chemical(s).}

\item{dtxsid}{A single or vector ofEPA's DSSTox Structure ID(s) 
(\url{https://comptox.epa.gov/dashboard})}

\item{casrn.vector}{A deprecated argument specifying a single or vector of 
Chemical Abstracts Service Registry 
Number(s) (CAS-RN) of desired chemical(s).}

\item{nomconc.vector}{For vector or single value, micromolar (uM = mol/L) nominal 
concentration (e.g. AC50 value)}

\item{this.well_number}{For single value, plate format default is 384, used
if is.na(tcdata)==TRUE. This value chooses default surface area settings for
\code{\link{armitage_estimate_sarea}} based on the number of wells per plate.}

\item{this.FBSf}{Fraction fetal bovine serum, must be entered by user.}

\item{tcdata}{A data.table with casrn, nomconc, MP, gkow, gkaw, gswat, sarea,
v_total, v_working. Otherwise supply single values to this.params (e.g., this.sarea,
this.v_total, etc.). Chemical parameters are taken from 
\code{\link{chem.physical_and_invitro.data}}.}

\item{user_assay_parameters}{option to fill in your own assay parameters (data table)}

\item{this.sarea}{Surface area per well (m^2)}

\item{this.v_total}{Total volume per well (uL)}

\item{this.v_working}{Working volume per well (uL)}

\item{this.cell_yield}{Number of cells per well}

\item{this.Tsys}{System temperature (degrees C)}

\item{this.Tref}{Reference temperature (degrees K)}

\item{this.option.kbsa2}{Use alternative bovine-serum-albumin partitioning
model}

\item{this.option.swat2}{Use alternative water solubility correction}

\item{this.option.kpl2}{Use alternative plastic-water partitioning model}

\item{this.option.bottom}{Include the bottom of the well in surface area calculation}

\item{this.pseudooct}{Pseudo-octanol cell storage lipid content}

\item{this.memblip}{Membrane lipid content of cells}

\item{this.nlom}{Structural protein content of cells}

\item{this.P_nlom}{Proportionality constant to octanol structural protein}

\item{this.P_dom}{Proportionality constant to dissolve organic material}

\item{this.P_cells}{Proportionality constant to octanol storage lipid}

\item{this.cell_pH}{7.4, pH of cell}

\item{this.Anionic_VF}{Anionic phospholipid fraction}

\item{this.A_Prop_acid}{Sorption to anionic lipids - acidic chemicals}

\item{this.A_Prop_base}{Sorption to anionic lipids - basic chemicals}

\item{this.Lyso_VF}{lysosome volume fraction}

\item{this.Lyso_Diam}{diameter of lysosome (500 nm)}

\item{this.Lyso_pH}{pH of lysosome (5.1)}

\item{this.csalt}{Ionic strength of buffer (M = mol/L)}

\item{this.celldensity}{Cell density kg/L, g/mL}

\item{this.cellmass}{Mass per cell, ng/cell}

\item{this.f_oc}{Everything assumed to be like proteins}

\item{this.conc_ser_alb}{Mass concentration of albumin in serum (g/L)}

\item{this.conc_ser_lip}{Mass concentration of lipids in serum (g/L)}

\item{this.Vdom}{0 ml, the volume of dissolved organic matter (DOM)}

\item{this.pH}{pH of cell culture}

\item{restrict.ion.partitioning}{FALSE, Should we restrict the chemical available to partition to only the neutral fraction?}

\item{surface.area.switch}{TRUE, automatically calculates surface area, switch to FALSE if user provided}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\value{
\tabular{lll}{
\strong{Param} \tab \strong{Description} \tab \strong{Units} \cr
casrn \tab Chemical Abstracts Service Registry Number \tab character \cr
nomconc \tab Nominal Concentration \tab uM=umol/L \cr       
well_number \tab Number of wells in plate (used to set default surface area) \tab unitless \cr   
sarea \tab Surface area of well \tab m^2 \cr         
v_total \tab Total volume of well \tab uL \cr       
v_working \tab Filled volume of well \tab uL \cr     
cell_yield \tab Number of cells \tab cells \cr    
assay_component_endpoint_name \tab link to invitro.assay.params table \tab character \cr  
gkow \tab The log10 octanol to water (PC) (logP)\tab log10 unitless ratio \cr          
logHenry \tab The log10 Henry's law constant '\tab log10 unitless ratio \cr      
gswat \tab The log10 water solubility (logWSol) \tab log10 mg/L \cr         
MP_C \tab The chemical compound melting point \tab degrees Celcius \cr  
MP_K \tab The chemical compound melting point \tab degrees Kelvin \cr        
MW \tab The chemical compound molecular weight \tab g/mol \cr            
gkaw \tab The air to water PC \tab unitless ratio \cr          
duow \tab internal energy of phase change for octanol-water \tab J/mol \cr          
duaw \tab internal energy of phase change for air-water \tab J/mol \cr          
gkmw \tab The log10 membrane to water PC \tab log10 unitless ratio \cr          
gkcw \tab The log10 cell/tissue to water PC \tab log10 unitless ratio\cr          
gkbsa \tab The log10 bovine serum albumin to water PC \tab log10 unitless ratio \cr         
gkpl \tab The log10 plastic to water PC \tab log10 m2/m2 \cr   
ksalt \tab Setschenow constant \tab L/mol \cr        
Tsys \tab System temperature \tab degrees C \cr          
Tref \tab Reference temperature\tab degrees K \cr          
option.kbsa2 \tab Use alternative bovine-serum-albumin partitioning model \tab logical \cr  
option.swat2 \tab Use alternative water solubility correction \tab logical \cr  
option.kpl2 \tab Use alternative plastic-water partitioning model \tab logical \cr 
FBSf \tab Fraction fetal bovine serum \tab unitless \cr          
pseudooct \tab Pseudo-octanol cell storage lipid content \tab \cr     
memblip \tab Membrane lipid content of cells \tab unitless \cr       
nlom \tab Structural protein content of cells \tab unitless \cr
P_nlom \tab Proportionality constant to octanol structural protein \tab unitless \cr   
P_dom \tab Proportionality constant to dissolved organic material (DOM) \tab unitless \cr         
P_cells \tab Proportionality constant to octanol storage lipid \tab unitless \cr      
Anionic_VF \tab Anionic phospholipid fraction \tab unitless \cr
A_Prop_acid \tab Sorption to anionic lipids - acidic chemicals \tab unitless \cr
A_Prop_base \tab Sorption to anionic lipids - basic chemicals \tab unitless \cr
Lyso_VF \tab Lysosome volume fraction \tab unitless \cr
Lyso_Diam \tab Diameter of lysosome \tab nm \cr
Lyso_pH \tab pH of lysosome \tab unitless \cr
csalt \tab Ionic strength of buffer \tab M=mol/L \cr
celldensity \tab Cell density \tab kg/L, g/mL \cr   
cellmass \tab Mass per cell \tab ng/cell \cr      
f_oc \tab Indicates fraction of dissolved organic matter to be treated like proteins \tab unitless \cr
cellwat \tab Fraction of the cell made up of water \tab unitless \cr       
Tcor \tab Temperature correction \tab \cr          
Vm \tab Volume of media \tab L \cr            
Vwell \tab Volume of medium (aqueous phase only) \tab L \cr         
Vair \tab Volume of head space \tab L \cr          
Vcells \tab Volume of cells/tissue\tab L \cr        
Valb \tab Volume of serum albumin \tab L \cr         
Vslip \tab Volume of serum lipids \tab L \cr         
Vdom \tab Volume of dissolved organic matter\tab L \cr          
F_ratio \tab Fugacity ratio \tab unitless \cr       
kmw \tab The membrane to water PC (i.e., 10^gkmow \tab unitless \cr           
kow \tab The octanol to water PC (i.e., 10^gkow) \tab unitless \cr           
kaw \tab The air to water PC (i.e., 10^gkaw) \tab unitless \cr           
swat \tab The intrinsic water solubility (i.e., 10^gswat) \tab mg/L \cr         
kpl \tab The plastic to water PC (i.e., 10^gkpl) \tab m3/m2 \cr           
kcw \tab The cell/tissue to water PC (i.e., 10^gkcw) \tab unitless \cr           
kbsa \tab The bovine serum albumin to water PC \tab unitless \cr          
swat_L \tab Water solubility limit used for Fugacity ratio calculation \tab \cr        
mtot \tab Total micromoles \tab umol \cr          
cwat \tab Total concentration in water \tab uM=umol/L \cr          
cwat_s \tab Dissolved concentration in water \tab uM=umol/L \cr        
csat \tab Is the solution saturated (1/0) \tab logical \cr         
activity \tab Chemical activity; indicates the potential for baseline toxicity to occur \tab \cr      
cair \tab Concentration in head space\tab uM=umol/L \cr          
calb \tab Concentration in serum albumin\tab uM=umol/L \cr          
cslip \tab Concentration in serum lipids\tab uM=umol/L \cr         
cdom \tab Concentration in dissolved organic matter\tab uM=umol/L \cr          
ccells \tab Concentration in cells\tab uM=umol/L \cr        
cplastic \tab Concentration in plastic\tab uM=umol/m^2 \cr      
mwat_s \tab Mass dissolved in water \tab umols \cr        
mair \tab Mass in air/head space \tab umols \cr          
mbsa \tab Mass bound to bovine serum albumin \tab umols \cr          
mslip \tab Mass bound to serum lipids \tab umols \cr        
mdom \tab Mass bound to dissolved organic matter \tab umols \cr          
mcells \tab Mass in cells \tab umols \cr        
mplastic \tab Mass bond to plastic \tab umols \cr      
mprecip \tab Mass precipitated out of solution \tab umols\cr       
xwat_s \tab Fraction dissolved in water \tab fraction \cr        
xair \tab Fraction in the air \tab fraction \cr          
xbsa \tab Fraction bound to bovine serum albumin \tab fraction \cr          
xslip \tab Fraction bound to serum lipids \tab fraction \cr         
xdom \tab Fraction bound to dissolved organic matter \tab fraction \cr          
xcells \tab Fraction within cells \tab fraction \cr        
xplastic \tab Fraction bound to plastic \tab fraction \cr     
xprecip \tab Fraction precipitated out of solution \tab fraction \cr       
eta_free \tab Effective availability ratio \tab fraction \cr      
\strong{cfree.invitro} \tab \strong{Free concentration in the in vitro media} (use for Honda1 and Honda2) \tab fraction \cr
}
}
\description{
Evaluate the Armitage model for chemical distributon \emph{in vitro}. Takes input
as data table or vectors of values. Outputs a data table. Updates over
the model published in Armitage et al. (2014) include binding to plastic walls
and lipid and protein compartments in cells.
}
\examples{

library(httk)

# Check to see if we have info on the chemical:
"80-05-7" \%in\% get_cheminfo()

#We do:
temp <- armitage_eval(casrn.vector = c("80-05-7", "81-81-2"), this.FBSf = 0.1,
this.well_number = 384, nomconc = 10)
print(temp$cfree.invitro)

\donttest{
# Check to see if we have info on the chemical:
"793-24-8" \%in\% get_cheminfo()

# Since we don't have any info, let's look up phys-chem from dashboard:
cheminfo <- data.frame(
  Compound="6-PPD",
  CASRN="793-24-8",
  DTXSID="DTXSID9025114",
  logP=4.27, 
  logHenry=log10(7.69e-8),
  logWSol=log10(1.58e-4),
  MP=	99.4,
  MW=268.404
  )
  
# Add the information to HTTK's database:
chem.physical_and_invitro.data <- add_chemtable(
 cheminfo,
 current.table=chem.physical_and_invitro.data,
 data.list=list(
 Compound="Compound",
 CAS="CASRN",
  DTXSID="DTXSID",
  MW="MW",
  logP="logP",
  logHenry="logHenry",
  logWSol="logWSol",
  MP="MP"),
  species="Human",
  reference="CompTox Dashboard 31921")

# Run the Armitage et al. (2014) model:
out <- armitage_eval(
  casrn.vector = "793-24-8", 
  this.FBSf = 0.1,
  this.well_number = 384, 
  nomconc = 10)
  
print(out)

}
}
\references{
\insertRef{armitage2014application}{httk}

\insertRef{honda2019using}{httk}
}
\author{
Greg Honda, Meredith Scherer adapted from code by James Armitage, Jon Arnot
}
