% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aylward2014}
\alias{aylward2014}
\alias{Aylward2014}
\title{Aylward et al. 2014}
\format{
data.frame
}
\source{
\insertRef{kapraun2022fetalmodel}{httk}
}
\usage{
aylward2014
}
\description{
Aylward et al. (2014) compiled measurements of the ratio of maternal to fetal 
cord blood chemical concentrations at birth for a range of chemicals with 
environmental routes of exposure, including bromodiphenyl ethers, fluorinated 
compounds, organochlorine pesticides, polyaromatic hydrocarbons, tobacco smoke 
components, and vitamins.
}
\references{
\insertRef{Aylward2014matfet}{httk}
}
\keyword{data}
