% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmiage}
\alias{bmiage}
\title{CDC BMI-for-age charts}
\format{
A data.table with 434 rows and 5 variables: \describe{
  \item{Sex}{Female or Male} \item{Agemos}{Age in months} \item{P5}{The 5th
  percentile BMI for the corresponding sex and age} \item{P85}{The 85th
  percentile BMI for the corresponding sex and age} \item{P95}{The 95th
  percentile BMI for the corresponding sex and age} }
}
\source{
\url{http://www.cdc.gov/growthcharts/data/zscore/bmiagerev.csv}
}
\usage{
bmiage
}
\description{
Charts giving the BMI-for-age percentiles for boys and girls ages 2-18
}
\details{
For children ages 2 to 18, weight class depends on the BMI-for-age percentile.
\describe{
\item{Underweight}{<5th percentile}
\item{Normal weight}{5th-85th percentile}
\item{Overweight}{85th-95th percentile}
\item{Obese}{>=95th percentile}
}
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability." Environment
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{data}
\keyword{httk-pop}
