% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ma.R
\name{calc_ma}
\alias{calc_ma}
\title{Calculate the membrane affinity}
\usage{
calc_ma(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  parameters = NULL,
  suppress.messages = FALSE,
  pfas.calibration = TRUE,
  chemdata = chem.physical_and_invitro.data
)
}
\arguments{
\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{dtxsid}{EPA's 'DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
 -- if parameters is not specified then the chemical must be identified by 
either CAS, name, or DTXSIDs}

\item{parameters}{Parameters from the appropriate parameterization function
for the model indicated by argument model}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{pfas.calibration}{Whether MA for chemicals in class PFAS should be
increased using the regression to the Droge (2019) dataset.}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\value{
A membrane:unbound fraction in plasma partition coefficient
}
\description{
Membrane affinity (MA) is the membrane:water partition coefficient. MA
chacterizes chemical partitioning into membranes formed
from neutral phospholipids (\ifelse{html}{\out{K<sub>nPL</sub>}}{\eqn{K_{nPL}}}). 
Pearce et al. (2017) compared five different methods for predicting 
membrane affinity using measured data for 59 compounds. The method of
Yun and Edgington (2013) was identified as the best: 
\ifelse{html}{\out{MA = 10^(1.294 + 0.304 * log<sub>10</sub>(P<sub>ow</sub>)}}{\deqn{MA = 10^(1.294 + 0.304 * log_{10}(P_{ow}))}}
}
\references{
\insertRef{pearce2017evaluation}{httk}

\insertRef{yun2013correlation}{httk}

\insertRef{droge2019pfasmembraneaffinity}{httk}
}
\author{
John Wambaugh
}
\keyword{parameters}
