% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_maternal_bw.R
\name{calc_maternal_bw}
\alias{calc_maternal_bw}
\title{Calculate maternal body weight}
\usage{
calc_maternal_bw(week = 12, chemdata = chem.physical_and_invitro.data)
}
\arguments{
\item{week}{Gestational week}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\value{
\item{BW}{Maternal Body Weight, kg.}
}
\description{
This function initializes the parameters needed in the functions
solve_fetal_pbtk by calling solve_pbtk and adding additional parameters.
}
\details{
BW <- params$pre_pregnant_BW + 
   params$BW_cubic_theta1 * tw + 
   params$BW_cubic_theta2 * tw^2 + 
   params$BW_cubic_theta3 * tw^3
}
\references{
\insertRef{kapraun2019empirical}{httk}
}
\author{
John Wambaugh
}
\keyword{Parameter}
