% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_total_clearance.R
\name{calc_total_clearance}
\alias{calc_total_clearance}
\title{Calculate the total plasma clearance.}
\usage{
calc_total_clearance(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  parameters = NULL,
  model = "3compartmentss",
  suppress.messages = FALSE,
  species = "Human",
  chemdata = chem.physical_and_invitro.data,
  ...
)
}
\arguments{
\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{dtxsid}{EPA's 'DSSTox Structure ID (https://comptox.epa.gov/dashboard)
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{parameters}{Chemical parameters from parameterize_steadystate
function, overrides chem.name and chem.cas.}

\item{model}{The model used to calculate total clearance (defaults to "3compartmentss")}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}

\item{...}{Additional parameters passed to parameterize function if 
parameters is NULL.}
}
\value{
\item{Total Clearance}{Units of L/h/kg BW.}
}
\description{
This function calculates the total clearance rate for a one compartment model
for plasma
where clearance is entirely due to metablism by the liver and glomerular
filtration in the kidneys, identical to clearance of three compartment
steady state model.
}
\examples{

calc_total_clearance(chem.name="Ibuprofen") 

}
\author{
John Wambaugh
}
\keyword{1compartment}
\keyword{Parameter}
