% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cas_id_check}
\alias{cas_id_check}
\title{CAS number format check function}
\usage{
cas_id_check(cas)
}
\arguments{
\item{cas}{A character string, or vector of character strings, indicating
CAS/CASRN number.}
}
\value{
Logical output (TRUE or FALSE) indicating whether the character string(s)
provided match the anticipated format for a CAS/CASRN chemical identifier.
}
\description{
This function checks whether the CAS/CARN chemical identifier follows the anticipated
format of XXXXXXX-YY-Z (i.e. 2-7 digits, 2 digits, and 1 digit, respectively).
}
