% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chem.invivo.PK.aggregate.data}
\alias{chem.invivo.PK.aggregate.data}
\title{Parameter Estimates from Wambaugh et al. (2018)}
\format{
data.frame
}
\source{
Wambaugh et al. 2018 Toxicological Sciences, in press
}
\usage{
chem.invivo.PK.aggregate.data
}
\description{
This table includes 1 and 2 compartment fits of plasma concentration vs time
data aggregated from chem.invivo.PK.data, performed in Wambaugh et al. 2018.
Data includes volume of distribution (Vdist, L/kg), elimination rate (kelim,
1/h), gut absorption rate (kgutabs, 1/h), fraction absorbed (Fgutabs), and
steady state concentration (Css, mg/L).
}
\author{
John Wambaugh
}
\keyword{data}
