% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dawson2021}
\alias{dawson2021}
\alias{Dawson2021}
\title{Dawson et al. 2021 data}
\format{
data.frame
}
\usage{
dawson2021
}
\description{
This table includes QSAR (Random Forest) model predicted values for unbound
fraction plasma protein (fup) and intrinsic hepatic clearance (clint) for a
subset of chemicals in the Tox21 library
(see \url{https://www.epa.gov/chemical-research/toxicology-testing-21st-century-tox21}).
}
\details{
Predictions were made with a set of Random Forest QSAR models,
as reported in Dawson et al. (2021).
}
\references{
\insertRef{dawson2021qsar}{httk}
}
\seealso{
\code{\link{load_dawson2021}}
}
\author{
Daniel E. Dawson
}
\keyword{data}
