% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dimitrijevic.IVD}
\alias{Dimitrijevic.IVD}
\title{Dimitrijevic et al. (2022)In Vitro Cellular and Nominal Concentration}
\format{
data.table and data.frame
}
\usage{
Dimitrijevic.IVD
}
\description{
Dimitrijevic et al. (2022)In Vitro Cellular and Nominal Concentration
}
\references{
\insertRef{dimitrijevic2022toward}{httk}
}
\author{
Jon Arnot
}
\keyword{datasets}
