% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT_age}
\alias{DT_age}
\title{Smoothed age distributions by race and gender.}
\format{
A data.table object with 9600 rows and 4 variables: \describe{
 \item{\code{gender}}{Gender: Male or Female}
 \item{\code{reth}}{Race/ethnicity} \item{\code{x}}{960 evenly-spaced grid points between 0 and 959 months}
 \item{\code{y}}{Smoothed density values evaluated at each \code{x} value}
 }
}
\usage{
DT_age
}
\description{
Distributions of ages in months, computed from NHANES data smoothed using
\code{\link[survey]{svysmooth}}, for each combination of race/ethnicity and gender.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability." Environment
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{data}
\keyword{httk-pop}
