% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtxsid_id_check}
\alias{dtxsid_id_check}
\title{DTXSID number format check function}
\usage{
dtxsid_id_check(dtxsid)
}
\arguments{
\item{dtxsid}{A character string, or vector of character strings, indicating
DTXSID number.}
}
\value{
Logical output (TRUE or FALSE) indicating whether the character string(s)
provided match the anticipated format for a DTXSID chemical identifier.
}
\description{
This function checks whether the DTXSID chemical identifier follows the anticipated
format of "DTXSID<uniqueID>".
}
