% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPA.ref}
\alias{EPA.ref}
\title{Reference for EPA Physico-Chemical Data}
\format{An object of class \code{character} of length 1.}
\source{
\url{https://comptox.epa.gov/dashboard}
}
\usage{
EPA.ref
}
\description{
The physico-chemical data in the chem.phys_and_invitro.data table are
obtained from EPA's Comptox Chemicals dashboard. This variable indicates
the date the Dashboard was accessed.
}
\author{
John Wambaugh
}
\keyword{datasets}
