% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gfr_ped.R
\name{estimate_gfr_ped}
\alias{estimate_gfr_ped}
\title{Predict GFR in children.}
\usage{
estimate_gfr_ped(BSA)
}
\arguments{
\item{BSA}{Vector of body surface areas in m^2.}
}
\value{
Vector of GFRs in mL/min/1.73m^2.
}
\description{
BSA-based equation from Johnson et al. 2006, Clin Pharmacokinet 45(9) 931-56. 
Used in Wetmore et al. 2014.
}
\references{
\insertRef{ring2017identifying}{httk}
}
\author{
Caroline Ring
}
\keyword{httk-pop}
