% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.seem}
\alias{example.seem}
\title{SEEM Example Data
We can grab SEEM daily intake rate predictions already in RData format from
https://github.com/HumanExposure/SEEM3RPackage/tree/main/SEEM3/data
Download the file Ring2018Preds.RData}
\format{
data.frame
}
\usage{
example.seem
}
\description{
We do not have the space to distribute all the SEEM predictions within
this R package, but we can give you our "Intro to IVIVE" example chemicals
}
\references{
\insertRef{Ring2017SEEM}{httk}
}
\keyword{data}
