% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.toxcast}
\alias{example.toxcast}
\title{ToxCast Example Data
The main page for the ToxCast data is here:
https://www.epa.gov/comptox-tools/exploring-toxcast-data
Most useful to us is a single file containing all the hits across all chemcials
and assays:
https://clowder.edap-cluster.com/datasets/6364026ee4b04f6bb1409eda?space=62bb560ee4b07abf29f88fef}
\format{
data.frame
}
\usage{
example.toxcast
}
\description{
As of November, 2022 the most recent version was 3.5 and was available as an
.Rdata file (invitrodb_3_5_mc5.Rdata)
}
\details{
Unfortunately for this vignette there are too many ToxCast data to fit into a
5mb R package. So we will subset to just the shemicals for the 
"Intro to IVIVE" vignette and distribute
only those data. In addition, out of 78 columns in the data, we will keep only 
eight.
}
\keyword{data}
