% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetalpcs}
\alias{fetalpcs}
\alias{fetalPCs}
\title{Fetal Partition Coefficients}
\format{
data.frame
}
\source{
\insertRef{kapraun2022fetalmodel}{httk}
}
\usage{
fetalpcs
}
\description{
Partition coefficients were measured for tissues, including placenta, in 
vitro by Csanady et al. (2002) for Bisphenol A and Diadzen. Curley et al. 
(1969) measured the concentration of a variety of pesticides in the cord 
blood of newborns and in the tissues of infants that were stillborn.
}
\details{
Three of the chemicals studied by Curley et al. (1969) were modeled by 
Weijs et al. (2013) using the same partition coefficients for mother and 
fetus. The values used represented "prior knowledge" summarizing the 
available literature.
}
\references{
\insertRef{Csanady2002fetalpc}{httk}
\insertRef{Curley1969fetalpc}{httk}
\insertRef{Weijs2013fetalpc}{httk}
}
\keyword{data}
