% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Frank2018invivo}
\alias{Frank2018invivo}
\title{Literature In Vivo Data on Doses Causing Neurological Effects}
\format{A data.frame containing 14 rows and 16 columns.}
\usage{
Frank2018invivo
}
\description{
Studies were selected from Table 1 in Mundy et al., 2015, as
the studies in that publication were cited as examples of
compounds with evidence for developmental neurotoxicity. There
were sufficient in vitro toxicokinetic data available for this
package for only 6 of the 42 chemicals.
}
\references{
Frank, Christopher L., et al. "Defining toxicological tipping points
in neuronal network development." Toxicology and Applied
Pharmacology 354 (2018): 81-93.

Mundy, William R., et al. "Expanding the test set: Chemicals with
potential to disrupt mammalian brain development." Neurotoxicology
and Teratology 52 (2015): 25-35.
}
\author{
Timothy J. Shafer
}
\keyword{data}
