% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frechen2021}
\alias{frechen2021}
\alias{pregnonpregaucs}
\alias{Frechen2021}
\title{Frechen et al. 2021}
\format{
data.frame
}
\source{
Kapraun et al. 2021 (submitted)
}
\usage{
pregnonpregaucs
}
\description{
Frechen et al. (2021) includes compiled literature descriptions of 
toxicokinetic summary statistics, including time-integrated plasma 
concentrations (area under the curve or AUC) for drugs administered to a 
sample of subjects including both pregnant and non-pregnant women. The 
circumstances of the dosing varied slightly between drugs and are summarized 
in the table.
}
\references{
\insertRef{Frechen2021matfetpbpk}{httk}
}
\keyword{data}
