% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_nhanes_params_pbtk_growthcurve.R
\name{gen_nhanes_params_pbtk_growthcurve}
\alias{gen_nhanes_params_pbtk_growthcurve}
\title{gen_nhanes_params_pbtk_growthcurve}
\usage{
gen_nhanes_params_pbtk_growthcurve(
  chem.cas.list = get_cheminfo(species = "Human", model = "pbtk")
)
}
\arguments{
\item{chem.cas.list}{A list of chemical CAS codes for determining chemical-specific
model parameter values. Without this option, the function generates parameter values
for all compounds with data available. WARNING: running this function without 
a short \code{chem.cas.list} can produce large file sizes and take up to twenty
minutes to complete. 

return}
}
\description{
A function which uses the same methods as \code{\link{parameterize_pbtk_growthcurve}}
to determine parameter values for every individual in the NHANES dataset.
}
\details{
Data are stored in \code{/httk/data/} according to the names of compiled parameters.
}
\examples{
\donttest{

gen_nhanes_params_growthcurve(chem.cas.list = c("80-05-7"))

}

}
\seealso{
\code{\link{parameterize_pbtk_growthcurve}}

\code{\link{get_compiled_param_timeseries}}

keywords
}
\author{
Colin Thomson

references 
insertRef
}
