% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_serum_creatinine.R
\name{gen_serum_creatinine}
\alias{gen_serum_creatinine}
\title{Predict GFR.}
\usage{
gen_serum_creatinine(serumcreat.dt)
}
\arguments{
\item{serumcreat.dt}{A data.table with columns \code{gender}, \code{reth}, 
\code{age_years}, \code{age_months}, \code{BSA_adj}.}
}
\value{
The same data.table with a \code{serum_creat} column added, containing 
spline-interpolated serum creatinine values.
}
\description{
Predict serum creatinine using smoothing splines and kernel density estimates of residual variability
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to 
environmental chemicals by simulating toxicokinetic variability." Environment 
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{httk-pop}
