% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caco2.R
\name{get_caco2}
\alias{get_caco2}
\title{Retrieve in vitro measured Caco-2 membrane permeabilit}
\usage{
get_caco2(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  Caco2.Pab.default = 1.6,
  suppress.messages = FALSE,
  chemdata = chem.physical_and_invitro.data
)
}
\arguments{
\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- the 
chemical must be identified by either CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  the 
chemical must be identified by either CAS, name, or DTXISD}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
-- the chemical must be identified by either CAS, name, or DTXSIDs}

\item{Caco2.Pab.default}{sets the default value for  Caco2.Pab if Caco2.Pab is 
unavailable.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\description{
This function checks for chemical-specific in vitro measurements of the 
Caco-2 membrane permeability 
in the \code{\link{chem.physical_and_invitro.data}} table. If no value is 
available argument \code{Caco2.Pab.default} is returned.
Anywhere that the values is reported by three numbers separated by a comma 
(this also happens for plasma protein binding) the three values are: median, 
lower 95 percent confidence intervals, upper 95 percent confidence interval. Unless you are 
doing monte carlo work it makes sense to ignore the second and third values.
}
\author{
John Wambaugh
}
\keyword{Parameter}
\keyword{oral_bioavailability}
