% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chem_id.R
\name{get_chem_id}
\alias{get_chem_id}
\title{Retrieve chemical identity from HTTK package}
\usage{
get_chem_id(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  chemdata = chem.physical_and_invitro.data
)
}
\arguments{
\item{chem.cas}{CAS regstry number}

\item{chem.name}{Chemical name}

\item{dtxsid}{DSSTox Substance identifier}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\value{
A list containing the following chemical identifiers:
\item{chem.cas}{CAS registry number}
\item{chem.name}{Name}
\item{dtxsid}{DTXSID}
}
\description{
Given one of chem.name, chem.cas (Chemical Abstract Service Registry Number),
or DTXSID (DSStox Substance Identifier \url{https://comptox.epa.gov/dashboard}) this
function checks if the chemical is available and, if so, returns all three
pieces of information.
}
\author{
John Wambaugh and Robert Pearce
}
\keyword{cheminformatics}
