% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compiled_param_timeseries.R
\name{get_compiled_param_timeseries}
\alias{get_compiled_param_timeseries}
\title{get_compiled_param_timeseries}
\usage{
get_compiled_param_timeseries(
  model,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  initial.params = NULL,
  initial.percentiles = NULL,
  start.age = 360,
  sim.duration = 10
)
}
\arguments{
\item{model}{The name of a model which can accept timeseries as forcing functions.}

\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- the 
chemical must be identified by either CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  the 
chemical must be identified by either CAS, name, or DTXISD}

\item{dtxsid}{EPA's 'DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
-- the chemical must be identified by either CAS, name, or DTXSIDs}

\item{initial.params}{The values for each parameter at the beginning of the simulation.
All compiled parameters should be present. The output of the parameterize_<model>
function is appropriate for \code{initial.params}.}

\item{initial.percentiles}{If \code{initial.params} are not provided, \code{initial.percentiles}
will designate a starting value for each parameter according to the corresponding 
percentile within the NHANES population. Values should be between zero and one. If 
neither \code{initial.params} nor \code{initial.percentiles} are provided, the initial
value for the parameter is taken to be the median of the NHANES population value.}

\item{start.age}{The age in months of the individual at the beginning of the simulation. 
Used for determining the percentile score of the initial parameter values when producing
the timeseries determining parameter changes.}

\item{sim.duration}{The length of the simulation in days. Equivalent to the \code{days}
parameter in \code{\link{solve_model}}.

param demographics}
}
\value{
A list of two-column matrices indexed by names of compiled parameters for
the designated model. The first column contains a list of times (in hours) and the 
second the total change in that parameter from the initial value.
}
\description{
The \code{deSolve} package uses timeseries as forcing functions. In lieu of hard-
coding time evolution of parameters into a model, these timeseries may be used
to change the value of parameters in time. The function \code{get_compiled_parm_timeseries}
queries the NHANES dataset and non-parametrically produces timeseries that 
preserve the percentile score of the given starting values.
}
\details{
For each time-dependent model, there should be a function (such as \code{\link{gen_nhanes_params_pbtk_growthcurve}})
that determines the model parameter values for each individual in the NHANES dataset.
The resulting value are used to form the non-parametric regression curve.
}
\examples{
\donttest{

params <- parameterize_pbtk_growthcurve(chem.name = 'Bisphenol A')
                      ts <- get_compiled_param_timeseries('pbtk_growthcurve',
                      chem.name = 'Bisphenol A',
                      initial.params = params,
                      start.age = 600, # age fifty
                      sim.duration = 365)

}

}
\seealso{
\code{\link{solve_pbtk_growthcurve}}

\code{\link{gen_nhanes_params_pbtk_growthcurve}}

keywords
}
\author{
Colin Thomson

references 
insertRef
}
