% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fabsgut.R
\name{get_fabsgut}
\alias{get_fabsgut}
\title{Retrieve or caclulate fraction of chemical absorbed from the gut}
\usage{
get_fabsgut(
  Params = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  Caco2.Pab.default = "1.6",
  Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE,
  overwrite.invivo = FALSE,
  keepit100 = FALSE,
  suppress.messages = FALSE
)
}
\arguments{
\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- the 
chemical must be identified by either CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  the 
chemical must be identified by either CAS, name, or DTXISD}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})
-- the chemical must be identified by either CAS, name, or DTXSIDs}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{default.to.human}{Substitutes missing rat values with human values if
true.}

\item{Caco2.Fgut}{= TRUE uses Caco2.Pab to calculate  fgut.oral, otherwise 
fgut.oral = \code{Fgut}.}

\item{Caco2.Fabs}{= TRUE uses Caco2.Pab to calculate
fabs.oral, otherwise fabs.oral = \code{Fabs}.}

\item{overwrite.invivo}{= TRUE overwrites Fabs and Fgut in vivo values from 
literature with Caco2 derived values if available.}

\item{keepit100}{TRUE overwrites Fabs and Fgut with 1 (i.e. 100 percent) 
regardless of other settings.}

\item{Caco2.default}{sets the default value for  Caco2.Pab if Caco2.Pab is 
unavailable.}
}
\description{
This function tetrieves or caclulates fraction of chemical absorbed from the gut.
We assume that systemic oral bioavailability (\ifelse{html}{\out{F<sub>bio</sub>}}{\eqn{F_{bio}}})
consists of three components: 
1) the fraction of chemical absorbed from intestinal lumen into enterocytes 
(\ifelse{html}{\out{F<sub>abs</sub>}}{\eqn{F_{abs}}}, 
2) the fraction surviving intestinal metabolism 
(\ifelse{html}{\out{F<sub>gut</sub>}}{\eqn{F_{gut}}}), and 
3) the fraction surviving first-pass hepatic metabolism 
(\ifelse{html}{\out{F<sub>hep</sub>}}{\eqn{F_{hep}}}). This function returns
\ifelse{html}{\out{F<sub>abs</sub>*F<sub>gut</sub>}}{\eqn{F_{abs}*F_{gut}}}
}
\details{
We model systemic oral bioavailability as 
\ifelse{html}{\out{F<sub>bio</sub>=F<sub>abs</sub>*F<sub>gut</sub>*F<sub>hep</sub>}}{\eqn{F_{bio}=F_{abs}*F_{gut}*F_{hep}}}.
\ifelse{html}{\out{F<sub>hep</sub>}}{\eqn{F_{hep}}}
is estimated from in vitro TK data using 
\code{\link{calc_hep_bioavailability}}.
If \ifelse{html}{\out{F<sub>bio</sub>}}{\eqn{F_{bio}}}
has been measured in vivo and is found in
table \code{\link{chem.physical_and_invitro.data)}} then we set 
\ifelse{html}{\out{F<sub>abs</sub>*F<sub>gut</sub>}}{\eqn{F_{abs}*F_{git}}} 
to the measured value divided by 
\ifelse{html}{\out{F<sub>hep</sub>}}{\eqn{F_{hep}}} 
Otherwise, if Caco2 membrane permeability data or predictions
are available \ifelse{html}{\out{F<sub>abs</sub>}}{\eqn{F_{abs}}} is estimated
using \code{\link{calc_fgut.oral}}.
Intrinsic hepatic metabolism is used to very roughly estimate
\ifelse{html}{\out{F<sub>gut</sub>}}{\eqn{F_{gut}}}
using \code{\link{calc_fgut.oral}}.
If argument keepit100 is used then there is complete absorption from the gut
(that is, \ifelse{html}{\out{F<sub>abs</sub>=F<sub>gut</sub>=1}}{\eqn{F_{abs}=F_{gut}=1}}).
}
\author{
Greg Honda and John Wambaugh
}
\keyword{Parameter}
\keyword{oral_bioavailability}
