% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rblood2plasma.R
\name{get_rblood2plasma}
\alias{get_rblood2plasma}
\title{Get ratio of the blood concentration to the plasma concentration.}
\usage{
get_rblood2plasma(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  suppress.messages = FALSE
)
}
\arguments{
\item{chem.name}{Chemical name (spaces and capitalization ignored) --  if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
 -- if parameters is not specified then the chemical must be identified by 
either CAS, name, or DTXSIDs}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{default.to.human}{Substitutes missing animal values with human values
if true.}

\item{suppress.messages}{Whether or not the output message is suppressed.}
}
\value{
A numeric value for the steady-state ratio of chemical concentration in blood
to plasma
}
\description{
This function attempts to retrieve a measured species- and chemical-specific 
blood:plasma concentration ratio from \code{\link{chem.phys_and_invitro.data}}.
If not available, Rblood2plasma is returned as NA.
}
\examples{

get_rblood2plasma(chem.name="Bisphenol A")
get_rblood2plasma(chem.name="Bisphenol A",species="Rat")

}
\seealso{
\code{\link{chem.phys_and_invitro.data}}
}
\author{
Robert Pearce, Annabel Meade
}
\keyword{Parameter}
\keyword{in-vitro}
