% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lit_cheminfo.R
\name{get_wetmore_cheminfo}
\alias{get_wetmore_cheminfo}
\title{Get literature Chemical Information. (deprecated).}
\usage{
get_wetmore_cheminfo(
  info = "CAS",
  species = "Human",
  suppress.messages = FALSE,
  chemdata = chem.physical_and_invitro.data
)
}
\arguments{
\item{info}{A single character vector (or collection of character vectors)
from
"Compound", "CAS", "MW", "Raw.Experimental.Percentage.Unbound",
"Entered.Experimental.Percentage.Unbound", "Fub", "source_PPB",
"Renal_Clearance", "Met_Stab", "Met_Stab_entered",
"r2", "p.val", "Concentration..uM.", "Css_lower_5th_perc.mg.L.", 
"Css_median_perc.mg.L.", "Css_upper_95th_perc.mg.L.",
"Css_lower_5th_perc.uM.","Css_median_perc.uM.","Css_upper_95th_perc.uM.",
and "Species".}

\item{species}{Species desired (either "Rat" or default "Human").}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\value{
\item{info}{Table/vector containing values specified in "info" for
valid chemicals.}
}
\description{
This function is included for backward compatibility. It calls
\code{\link{get_lit_cheminfo}} which
provides the information specified in "info=" for all
chemicals with data from the Wetmore et al. (2012) and (2013) publications
and other literature.
}
\examples{

get_lit_cheminfo()
get_lit_cheminfo(info=c('CAS','MW'))

}
\references{
\insertRef{wetmore2012integration}{httk}

\insertRef{wetmore2013relative}{httk}

\insertRef{wetmore2015incorporating}{httk}
}
\author{
John Wambaugh
}
\keyword{Literature}
\keyword{Retrieval}
