% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hct_spline}
\alias{hct_spline}
\title{Smoothing splines for log hematocrit vs. age in months}
\format{
A list with 10 elements, named for unique combinations of gender
 ("Male" and "Female") and NHANES race/ethnicity categories ("Mexican
 American", "Non-Hispanic White", "Non-Hispanic Black", "Other", and "Other
 Hispanic"). Each list element consists of an object of class
 \code{smooth.spline.fit} (output of \code{\link[stats]{smooth.spline}})
 which can be used with method \code{\link[stats]{predict.smooth.spline}} to
 predict natural-log hematocrit from age in months. The smoothing splines
 were fitted to NHANES data for each combination of NHANES gender and
 race/ethnicity categories: natural log hematocrit vs. age in months,
 weighted using the NHANES sample weights.
}
\usage{
hct_spline
}
\description{
Smoothing splines fitted to NHANES hematocrit and age data by race/ethnicity
and gender.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability." Environment
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{data}
\keyword{httk-pop}
