% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{honda2023.qspr}
\alias{honda2023.qspr}
\title{<Honda 2023 QSPR Title>
<DESCRIPTION NEEDED>}
\format{
An object of class \code{data.frame} with 14033 rows and 5 columns.
}
\usage{
honda2023.qspr
}
\description{
<Honda 2023 QSPR Title>
<DESCRIPTION NEEDED>
}
\keyword{datasets}
