% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{httk.performance}
\alias{httk.performance}
\title{<HTTK Benchmark Title>
<DESCRIPTION NEEDED>}
\format{
An object of class \code{data.frame} with 22 rows and 18 columns.
}
\usage{
httk.performance
}
\description{
<HTTK Benchmark Title>
<DESCRIPTION NEEDED>
}
\keyword{datasets}
