% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk_vignettes.R
\name{httk_vignettes}
\alias{httk_vignettes}
\title{Interact with HTTK vignettes}
\usage{
httk_vignettes(vignette = NULL, ...)
}
\arguments{
\item{vignette}{The name of a vignette to be displayed.}

\item{...}{Additional arguments to function \code{\link{vignette}}}
}
\description{
This function lists the available vignettes, including those from the
opetional httkexamples package.
}
\examples{
calc_analytic_css(chem.name='Bisphenol-A',output.units='mg/L',
                 model='3compartment',concentration='blood')


}
\seealso{
\code{\link{vignette}}
}
\author{
John Wambaugh
}
\keyword{vignettes}
