% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httkpop_generate_all_nhanes.R
\name{httkpop_generate_all_nhanes}
\alias{httkpop_generate_all_nhanes}
\title{httkpop_generate but for full NHANES population}
\usage{
httkpop_generate_all_nhanes()
}
\value{
A `data.table` as for [httkpop_generate()]
}
\description{
Returns results as for [httkpop_generate()], but using full NHANES population
}
\details{
This is equivalent to running [httkpop_generate()] but without resampling
}
\author{
Colin Thomson
}
