% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httkpop_mc.R
\name{httkpop_mc}
\alias{httkpop_mc}
\title{httk-pop: Correlated human physiological parameter Monte Carlo}
\usage{
httkpop_mc(model, samples = 1000, httkpop.dt = NULL, ...)
}
\arguments{
\item{model}{One of the HTTK models: "1compartment", "3compartmentss",
"3compartment", or "pbtk".}

\item{samples}{The number of Monte Carlo samples to use (can often think of these
as separate individuals)}

\item{httkpop.dt}{A data table generated by \code{\link{httkpop_generate}}.
This defaults to NULL, in which case \code{\link{httkpop_generate}} is 
called to generate this table.}

\item{...}{Additional arugments passed on to \code{\link{httkpop_generate}}.}
}
\value{
A data.table with a row for each individual in the sample and a column for
each parameter in the model.
}
\description{
This is the core function for httk-pop correlated human physiological 
variability simulation as described by Ring et al. (2017) 
(\doi{10.1016/j.envint.2017.06.004}). This functions 
takes the data table of population biometrics (one individual per row)
generated by \code{\link{httkpop_generate}}, and converts it
to the corresponding table of HTTK model parameters for a specified HTTK 
model.
}
\details{
The Monte Carlo methods used here were recently updated and described by
Breen et al. (submitted).
}
\examples{

set.seed(42)
indiv_examp <- httkpop_generate(method="d", nsamp=10)

httk_param <- httkpop_mc(httkpop.dt=indiv_examp, 
                        samples=10,
                        model="1compartment")

}
\references{
\insertRef{ring2017identifying}{httk} 

\insertRef{breen2022simulating}{httk} 

\insertRef{rowland1973clearance}{httk}
}
\author{
Caroline Ring and John Wambaugh
}
\keyword{httk-pop}
\keyword{monte-carlo}
