% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invitro.assay.params}
\alias{invitro.assay.params}
\title{ToxCast In Vitro Assay Descriptors}
\format{
data.table and data.frame
}
\usage{
invitro.assay.params
}
\description{
ToxCast In Vitro Assay Descriptors
}
\author{
Madison Feshuk
}
\keyword{datasets}
