% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IVD_Manuscript_data}
\alias{IVD_Manuscript_data}
\title{Literature Measurments of In Vitro Cellular and Nominal Concentration}
\format{
data.table and data.frame
}
\usage{
IVD_Manuscript_data
}
\description{
Literature Measurments of In Vitro Cellular and Nominal Concentration
}
\author{
Meredith Scherer
}
\keyword{datasets}
