% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kapraun2019}
\alias{kapraun2019}
\alias{Kapraun2019}
\title{Kapraun et al. 2019 data}
\format{
list
}
\source{
\insertRef{kapraun2022fetalmodel}{httk}
}
\usage{
kapraun2019
}
\description{
A list object containing time-varying parameters for the human maternal-fetal
HTTK model. List elements contain scalar coefficients for the polynomial,
logistic, Gompertz, and other functions of time describing blood flow rates,
tissue volumes, hematocrits, and other anatomical/physiological quantities
that change in the human mother and her fetus during pregnancy and gestation.
}
\references{
\insertRef{kapraun2019empirical}{httk}
}
\author{
Dustin F. Kapraun
}
\keyword{data}
