% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kde_centers}
\alias{kde_centers}
\title{Data points for kernel density estimation of residual variability of NHANES
quantities}
\format{
A data.table with 17494 rows and 7 columns: \describe{
  \item{g}{Gender, "Male" or "Female".} \item{r}{NHANES race/ethnicity
  category: "("Mexican American", "Non-Hispanic White", "Non-Hispanic Black",
  "Other", and "Other Hispanic")} \item{seqn}{NHANES unique identifier
  (five-digit integer code) identifying a unique NHANES respondent}
  \item{logwresid}{Weight residuals: natural-log NHANES weight (\code{bmxwt})
  - prediction of weight smoothing spline (\code{\link{weight_spline}}) for
  the individual identified by \code{seqn}} \item{loghresid}{Height
  residuals: natural-log NHANES height (\code{bmxwt}) - prediction of height
  smoothing spline (\code{\link{height_spline}}) for the individual
  identified by \code{seqn}} \item{loghctresid}{Hematocrit residuals:
  natural-log NHANES hematocrit (\code{lbxhct}) - prediction of hematocrit
  smoothing spline (\code{\link{hct_spline}}) for the individual identified
  by \code{seqn}} \item{logscresid}{Serum creatinine residuals: natural-log
  NHANES serum creatinine (\code{lbxscr}) - prediction of serum creatinine
  smoothing spline (\code{\link{scr_spline}}) for the individual identified
  by \code{seqn}} }
}
\source{
Spline fits to the data in \code{\link{nhanes_mec_svy}}
}
\usage{
kde_centers
}
\description{
In virtual-individuals mode, HTTK-Pop generates height, weight, serum
creatinine, and hematocrit values for simulated individuals by first using
the gender- and race/ethnicity-specific fitted smooothing splines for each of
these quantities to predict a value based on age. Then, HTTK-Pop adds
residual variability to the spline-predicted value. Residual variability is
estimated using a kernel-density estimate (KDE), which is determined in the
following way. First, the fitted splines are used to predict a value for each
NHANES respondent. Then, residuals are calculated as the difference between
the NHANES measured value and the spline-predicted value. Then, Gaussian
kernel distributions are placed centered on each of these residuals. (To
account for sample weights, centers are effectively repeated a number of
times corresponding to their relative weight.) The Gaussian kernel's
bandwidth (variance) is determined using the plug-in estimator from
\code{\link[ks]{Hpi}}. The sum of all the kernel distributions approximates
the 2-D distribution of residuals.
}
\details{
To sample from the resulting KDE for each quantity, the following algorithm
is used. First, draw a random sample with replacement from the "kernel
centers", i.e. the original vector of residuals. (Use the NHANES sample
weights to determine the probability of choosing each center.) Then, for each
sampled center, draw a random value from a normal distribution whose mean is
the sampled center, and whose standard deviation is the optimal KDE
bandwidth.

This data table contains the "kernel centers" -- the original vector of
residuals -- for each quantity whose residual variability is modeled using
KDE. It is used to generate samples from the KDE for residual variability,
along with the optimal bandwidth for each quantity (see \code{\link{hw_H}},
\code{\link{hct_h}}, and \code{\link{scr_h}}). (The necessary NHANES sample
weights are accessed by using column \code{seqn} to look up the sample
weights in the NHANES data table \code{\link{nhanes_mec_svy}}. NA values are
removed before sampling from each vector of centers.
}
\keyword{datasets}
