% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List all available HTTK models}
\usage{
list_models(names.only = FALSE)
}
\arguments{
\item{names.only}{If true, only return the model names}
}
\value{
Describes (or lists) available HTTK models
}
\description{
List all available HTTK models
}
\author{
John Wambaugh
}
