% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfas.clearance}
\alias{pfas.clearance}
\title{Interspecies In vivo Clearance Data for PFAS}
\format{
data.frame
}
\usage{
pfas.clearance
}
\description{

}
\details{
The data.frame contains the following columns:
\tabular{ll}{
\strong{Column Name} \tab \strong{Description} \cr
 DTXSID \tab CompTox Chemicals Dashboard substance identifier \cr                      
 Species \tab Species for which the clearance was calculated \cr
 Sex \tab Sex for which the clearance was calculated \cr
 HalfLifeHours \tab Half-life in hours \cr
 HlReference \tab Reference(s) for half-life \cr  
 VdLpkgbw  \tab Volume of distribution in L/kg body weight \cr
 VdReference \tab Reference for volume of distribution \cr
 Kelimph  \tab Elimination rate in 1/hour \cr
 ClLphpkgbw \tab Total clearance in L/h/kg body weight \cr
}
}
\references{
\insertRef{dawson2023machine}{httk}

\insertRef{fenton2021per}{httk}

\insertRef{pizzurro2019interspecies}{httk}

\insertRef{huang2019toxicokinetics}{httk}

\insertRef{dzierlenga2020toxicokinetics}{httk}

\insertRef{lau2020pharmacokinetic}{httk}

\insertRef{lou2009modeling}{httk}

\insertRef{tatum2011comparative}{httk}
}
\keyword{data}
