% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharma}
\alias{pharma}
\title{DRUGS|NORMAN: Pharmaceutical List with EU, Swiss, US Consumption Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 14 rows and 954 columns.
}
\source{
\url{https://comptox.epa.gov/dashboard/chemical_lists/swisspharma}
}
\usage{
pharma
}
\description{
SWISSPHARMA is a list of pharmaceuticals with consumption data from
Switzerland, France, Germany and the USA, used for a suspect
screening/exposure modelling approach described in
Singer et al 2016, DOI: 10.1021/acs.est.5b03332. The original data is
available on the NORMAN Suspect List Exchange.
}
\references{
Wambaugh et al. (2019) "Assessing Toxicokinetic Uncertainty and
Variability in Risk Prioritization", Toxicological Sciences, 172(2), 235-251.
}
\keyword{data}
