% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physiology.data}
\alias{physiology.data}
\title{Species-specific physiology parameters}
\format{
A data.frame containing 18 rows and 7 columns.
}
\usage{
physiology.data
}
\description{
This data set contains values from Davies and Morris (1993) necessary to
paramaterize a toxicokinetic model for human, mouse, rat, dog, or rabbit.
The temperature for each species are taken from Reece (2015),
Jordon (1995), and Stammers (1926). Mean residence time for the small 
intestine is from Grandoni et al. (2019). Human small intestine radius is
from Yu et al. (1999). Rat small intestine radius is from Griffin and
O'Driscoll (2008).
}
\references{
\insertRef{davies1993physiological}{httk}

\insertRef{brown1997physiological}{httk}

\insertRef{birnbaum1994physiological}{httk}

\insertRef{reece201514}{httk}

\insertRef{stammers1926blood}{httk}

\insertRef{jordan1995temperature}{httk}

\insertRef{grandoni2019building}{httk}

\insertRef{griffin2008models}{httk}
}
\author{
John Wambaugh and Nisha Sipes
}
\keyword{data}
