% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie_chart.R
\name{piechart}
\alias{piechart}
\title{Create a pie chart for use in visualizing in vitro distribution}
\usage{
piechart(input_table)
}
\arguments{
\item{input_table}{Object of class data.frame with only the following columns:
\tabular{ccc}{
\strong{Column Name} \tab \strong{Description} \cr
compartment \tab Compartment or location of chemical \cr
fraction \tab Fraction of chemical in the given compartment \cr
}}
}
\value{
A pie chart created in ggplot.
}
\description{
This function accepts a list of compartments and fractions and generates a pie chart. The chart is designed for use with in vitro distribution models with the goal of visualizing the distribution of a chemical.
}
\examples{
#import ggplot2 (required)
library("ggplot2")
#create an input table with two columns: compartment and fraction
input_table <- data.frame(compartment = c("Free in medium", "Bound in plasma", "Associated with cells", "In headspace", "Soaked to well plastic"), fraction=c(0.1, 0.240597245, 0.600, 0.048, 0.01140276))
#call the piechart function
piechart(input_table)
}
\author{
Meredith Scherer
}
