\name{PK.physiology.data}
\docType{data}
\alias{PK.physiology.data}
\title{Species-specific physiology parameters}
\description{
This data set contains values from Davies and Morris (1993) necessary
to paramaterize a toxicokinetic model for human, mouse, rat, dog, or 
rabbit.  The temperature for each species are taken from
Robertshaw et al. (1926) and Gordon (1993)  
}
\usage{PK.physiology.data}
\format{A data.frame containing 11 rows and 7 columns.}
\source{Wambaugh et al. (2014), in preparation}
\references{
Davies, B. and Morris, T. (1993). Physiological Parameters in Laboratory Animals and Humans. Pharmaceutical Research 10(7), 1093-1095, 10.1023/a:1018943613122.

Robertshaw, D., Temperature Regulation and Thermal Environment, in Dukes' Physiology of Domestic Animals, 12th ed., Reece W.O., Ed. Copyright 2004 by Cornell University.
 Stammers (1926) The blood count and body temperature in normal rats
 Gordon (1993) Temperature Regulation in Laboratory Rodents
 }
\author{
John Wambaugh and Nisha Sipes
}
\keyword{ datasets }
