% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pradeep2020}
\alias{pradeep2020}
\alias{Pradeep2020}
\title{Pradeep et al. 2020}
\format{
data.frame
}
\usage{
pradeep2020
}
\description{
This table includes Support Vector Machine and Random Forest model predicted
values for unbound fraction plasma protein (fup) and intrinsic hepatic
clearance (clint) values for a subset of chemicals in the Tox21 library
(see \url{https://www.epa.gov/chemical-research/toxicology-testing-21st-century-tox21}).
}
\details{
Prediction were made with Support Vector Machine and Random Forest models,
as reported in Pradeep et al. (2020).
}
\references{
\insertRef{pradeep2020chemstr}{httk}
}
\seealso{
\code{\link{load_pradeep2020}}
}
\keyword{data}
