% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_partitioning_schmitt.R
\name{predict_partitioning_schmitt}
\alias{predict_partitioning_schmitt}
\title{Predict partition coefficients using the method from Schmitt (2008).}
\usage{
predict_partitioning_schmitt(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  species = "Human",
  model = "pbtk",
  default.to.human = FALSE,
  parameters = NULL,
  alpha = 0.001,
  adjusted.Funbound.plasma = TRUE,
  regression = TRUE,
  regression.list = c("brain", "adipose", "gut", "heart", "kidney", "liver", "lung",
    "muscle", "skin", "spleen", "bone"),
  tissues = NULL,
  minimum.Funbound.plasma = 1e-04,
  suppress.messages = FALSE,
  chemdata = chem.physical_and_invitro.data
)
}
\arguments{
\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{model}{Model for which partition coefficients are neeeded (for example,
"pbtk", "3compartment")}

\item{default.to.human}{Substitutes missing animal values with human values
if true (hepatic intrinsic clearance or fraction of unbound plasma).}

\item{parameters}{Chemical parameters from \code{\link{parameterize_schmitt}}
overrides chem.name, dtxsid, and chem.cas.}

\item{alpha}{Ratio of Distribution coefficient D of totally charged species
and that of the neutral form}

\item{adjusted.Funbound.plasma}{Whether or not to use Funbound.plasma
adjustment.}

\item{regression}{Whether or not to use the regressions.  Regressions are
used by default.}

\item{regression.list}{Tissues to use regressions on.}

\item{tissues}{Vector of desired partition coefficients.  Returns all by
default.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}
}
\value{
Returns tissue to unbound plasma partition coefficients for each
tissue.
}
\description{
This function implements the method from Schmitt (2008) for predicting the 
tissue to unbound plasma partition coefficients for the tissues contained 
in the \code{\link{tissue.data}} table. The method has been modified
by Pearce et al. (2017) based on an evaluation using in vivo measured 
partition coefficients.

To understand this method, it is 
important to recognize that in  a given media the fraction unbound in that 
media is inverse of the media:water partition coefficient. 
In Schmitt's model, each tissue is composed of cells and
interstitium, with each cell consisting of neutral lipid,
neutral phospholipid, water, protein, and acidic phospholipid.
Each tissue cell is defined as the sum of separate
compartments for each constituent, all of which partition
with a shared water compartment. The partitioning between
the cell components and cell water is compound specific
and determined by log Pow (in neutral lipid partitioning),
membrane affinity (phospholipid and protein partitioning),
and pKa (neutral lipid and acidic phospholipid partitioning).
For a given compound the partitioning into each
component is identical across tissues. Thus the differences
among tissues are driven by their composition, that is, the
varying volumes of components such as neutral lipid.
However, pH differences across tissues also determine
small differences in partitioning between cell and plasma
water. The fup is used as the plasma water to total plasma
partition coefficient and to approximate the partitioning
between interstitial protein and water.

A regression is used to predict membrane affinity when measured values are 
not available (\code{\link{calc_ma}}).  The
regressions for correcting each tissue are performed on tissue plasma
partition coefficients (Ktissue2pu * Funbound.plasma) calculated with the
corrected Funbound.plasma value and divided by this value to get Ktissue2pu.
Thus the regressions should be used with the corrected Funbound.plasma.

A separate regression is used when adjusted.Funbound.plasma is FALSE.

The red blood cell regression can be used but is not by default because of
the span of the data used for evaluation, reducing confidence in the 
regression for higher and lower predicted values.

Human tissue volumes are used for species other than Rat.
}
\examples{

library(httk)

# Predict the partition coefficients by chemical id:
PCs1 <- predict_partitioning_schmitt(chem.name='ibuprofen')

# Create a list of parameters (that you can potentially change):
p <- parameterize_schmitt(chem.name="ibuprofen")

# Predict the partition coefficients using a list of parameters:
PCs2 <- predict_partitioning_schmitt(parameters = p)

# Check that all the parameter values are the same:
all(unlist(PCs1)==unlist(PCs2))

# Predict partition coefficients without using Pearce et al. (2017) calibrations:
PCs3 <- predict_partitioning_schmitt(chem.name='ibuprofen',regression=FALSE)

# Lump the tissues into the compartments for model "pbtk":
lump_tissues(PCs1)

# Lump the tissues into a single volume of distribution
calc_vdist(parameters=c(p,PCs1))

}
\references{
\insertRef{schmitt2008general}{httk}

\insertRef{birnbaum1994physiological}{httk}

\insertRef{pearce2017evaluation}{httk}

\insertRef{yun2013correlation}{httk}
}
\seealso{
\code{\link{parameterize_schmitt}}

\code{\link{tissue.data}}

\code{\link{calc_ma}}
}
\author{
Robert Pearce
}
\keyword{Parameter}
