% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pregnonpregaucs}
\alias{pregnonpregaucs}
\title{AUCs for Pregnant and Non-Pregnant Women}
\format{
data.frame
}
\source{
\insertRef{kapraun2022fetalmodel}{httk}
}
\usage{
pregnonpregaucs
}
\description{
Dallmann et al. (2018) includes compiled literature descriptions of 
toxicokinetic summary statistics, including time-integrated plasma 
concentrations (area under the curve or AUC) for drugs administered to a 
sample of subjects including both pregnant and non-pregnant women. The 
circumstances of the dosing varied slightly between drugs and are summarized 
in the table.
}
\references{
\insertRef{dallmann2018pregpbtk}{httk}
}
\keyword{data}
