% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_invitrouv_1comp.R
\name{propagate_invitrouv_1comp}
\alias{propagate_invitrouv_1comp}
\title{Propagates uncertainty and variability in in vitro HTTK data into one
compartment model parameters}
\usage{
propagate_invitrouv_1comp(
  parameters.dt,
  chemdata = chem.physical_and_invitro.data,
  ...
)
}
\arguments{
\item{parameters.dt}{The data table of parameters being used by the Monte
Carlo sampler}

\item{chemdata}{A data frame with physicochemical data following the exact
structure of httk's chem.physical_and_invitro.data data frame; the data frame 
must be either the original chem.physical_and_invitro.data data frame or the 
original chem.physical_and_invitro.data data frame with additional
rows of chemicals (if the user wanted to add chemicals to the list). All 
columns must remain and be in the same order as the original data frame.}

\item{...}{Additional arguments passed to \code{\link{calc_elimination_rate}}}
}
\value{
A data.table whose columns are the parameters of the HTTK model
 specified in \code{model}.
}
\description{
Propagates uncertainty and variability in in vitro HTTK data into one
compartment model parameters
}
\author{
John Wambaugh
}
\keyword{1compartment}
\keyword{monte-carlo}
