% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_left_censored_norm.R
\name{r_left_censored_norm}
\alias{r_left_censored_norm}
\title{Returns draws from a normal distribution with a lower censoring limit of lod
(limit of detection)}
\usage{
r_left_censored_norm(n, mean = 0, sd = 1, lod = 0.005, lower = 0, upper = 1)
}
\arguments{
\item{n}{Number of samples to take}

\item{mean}{Mean of censored distribution. Default 0.}

\item{sd}{Standard deviation of censored distribution. Default 1.}

\item{lod}{Bound below which to censor. Default 0.005.}

\item{lower}{Lower bound on censored distribution. Default 0.}

\item{upper}{Upper bound on censored distribution. Default 1.}
}
\value{
A vector of samples from the specified censored distribution.
}
\description{
Returns draws from a normal distribution with a lower censoring limit of lod
(limit of detection)
}
