% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Scherer2025.IVD}
\alias{Scherer2025.IVD}
\title{Literature Measurements of In Vitro Cellular and Nominal Concentration}
\format{
data.table and data.frame
}
\usage{
Scherer2025.IVD
}
\description{
Literature Measurements of In Vitro Cellular and Nominal Concentration
}
\author{
Meredith Scherer
}
\keyword{datasets}
