% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scr_h}
\alias{scr_h}
\title{Optimal bandwidth for kernel density estimate of serum creatinine residual
variability}
\format{
A list with 10 elements, named for unique combinations of NHANES
  gender categories ("Male" and "Female") and NHANES race/ethnicity
  categories ("Mexican American", "Non-Hispanic White", "Non-Hispanic Black",
  "Other", and "Other Hispanic"). Each list element is a numeric scalar
  giving the standard deviation that is the optimal KDE bandwidth for
  estimating the distribution of residual variability in (natural-log) serum
  creatinine residuals.
}
\source{
The output of \code{\link[ks]{hpi}} called on the vector of serum
  creatinine residuals (in \link{kde_centers}, column \code{logscresid}, with
  NAs removed)
}
\usage{
scr_h
}
\description{
In virtual-individuals mode, HTTK-Pop generates serum creatinine for
simulated individuals, by first using the gender- and race/ethnicity-specific
fitted smooothing splines for serum creatinine to predict values based on
age. Then, HTTK-Pop adds residual variability to the spline-predicted values.
Residual variability is estimated using a kernel-density estimate (KDE),
which is determined in the following way. First, the fitted splines are used
to predict (natural-log) serum creatinine values for each NHANES respondent.
Then, residuals are calculated as the difference between the natural-log
NHANES measured value and the spline-predicted value. Then, a Gaussian kernel
distribution is placed centered on each residual. (To account for sample
weights, centers are effectively repeated a number of times corresponding to
their relative weight.) The Gaussian kernel's bandwidth (standard deviation)
is determined using the plug-in estimator from \code{\link[ks]{hpi}}. The sum
of all the kernel distributions approximates the distribution of residuals.
}
\details{
To sample from the resulting KDE, the following algorithm is used. First,
draw a random sample with replacement from the "kernel centers", i.e. the
original vector of residuals. (Use the NHANES sample weights to determine the
probability of choosing each center.) Then, for each sampled center (serum
creatinine residual), draw a random value from a normal distribution whose
mean is the sampled serum creatinine residual, and whose standard deviation
is the optimal KDE bandwidth.
}
\keyword{datasets}
