% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipes2017}
\alias{sipes2017}
\title{<DESCRIPTION NEEDED>}
\format{
An object of class \code{data.frame} with 8758 rows and 3 columns.
}
\usage{
sipes2017
}
\description{
<DESCRIPTION NEEDED>
}
\keyword{datasets}
