% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spline_heightweight}
\alias{spline_heightweight}
\title{Smoothing splines for log height vs. age and log body weight vs. age, along
with 2-D KDE residuals, by race and gender.}
\format{
A data.table with 6 variables: \describe{ \item{\code{g}}{Gender: Male
 or Female} \item{\code{r}}{Race/ethnicity: Mexican American, Other Hispanic,
 Non-Hispanic White, Non-Hispanic Black, Other} \item{\code{height_spline}}{A
 list of smooth.spline objects, each giving a smoothed relationship between
 log height in cm and age in months} \item{\code{weight_spline}}{A list of
 smooth.spline objects, each giving a smoothed relationship between log body
 weight in kg and age in months} \item{\code{hw_kde}}{A list of kde objects;
 each is a 2-D KDE of the distribution of log height and log body weight
 residuals about the smoothing splines.} }
}
\usage{
spline_heightweight
}
\description{
#'Smoothing splines and KDE fits to joint distribution of height and weight
residuals pre-calculated from NHANES height, weight, and age data by
race/ethnicity and gender.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability." Environment
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{data}
\keyword{httk-pop}
