% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spline_serumcreat}
\alias{spline_serumcreat}
\title{Smoothing splines for log serum creatinine vs. age in months, along with KDE
residuals, by race and gender.}
\format{A data.table with 6 variables: \describe{
\item{\code{gender}}{Gender:
 Male or Female}
 \item{\code{reth}}{Race/ethnicity: Mexican American, Other
 Hispanic, Non-Hispanic White, Non-Hispanic Black, Other}
 \item{\code{sc_spline}}{A list of smooth.spline objects, each giving a
 smoothed relationship between log serum creatinine and age in months}
 \item{\code{sc_kde}}{A list of kde
 objects; each is a KDE of the distribution of residuals about the smoothing
 spline.}
  }}
\usage{
spline_serumcreat
}
\description{
#'Smoothing splines and KDE residuals pre-calculated from NHANES serum creatinine and
age data by race/ethnicity and gender.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability." Environment
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{data}
\keyword{httk-pop}
