% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supptab1_Linakis2020}
\alias{supptab1_Linakis2020}
\title{Supplementary output from Linakis 2020 vignette analysis.}
\format{
A data.frame containing x rows and y columns.
}
\source{
Matt Linakis
}
\usage{
supptab1_Linakis2020
}
\description{
Supplementary output from Linakis 2020 vignette analysis.
}
\references{
DSStox database (http:// www.epa.gov/ncct/dsstox
}
\author{
Matt Linakis
}
\keyword{data}
