% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thyroid.ac50s}
\alias{thyroid.ac50s}
\title{ToxCast thyroid-related bioactivity data}
\format{
data.table and data.frame
}
\usage{
thyroid.ac50s
}
\description{
Truong et al. 2025 uses ToxCast data for 4 thyroid-related assay endpoints
concerning inhibition of deiodinases ("DIO1", "DIO2", "DIO3", and "IYD") and 
identified 120 priority chemicals with activity for at least one deiodinase. 
These 120 chemicals were curated after assessment for target selectivity and 
assay interference.
}
\details{
The AC50s (in uM) for each of the 120 chemicals were retrieved from ToxCast 
invitrodb v3.5 and are used in the "Full Human Gestational IVIVE" vignette.
}
\keyword{data}
