\name{tissue.data}
\docType{data}
\alias{tissue.data}
\title{Tissue composition and species-specific physiology parameters}
\description{
This data set contains values from Schmitt (2008) describing the composition
of specific tissues and from Birnbaum et al. (1994) describing
volumes of and blood flows to thos tissues, allowing parameterization of 
toxicokinetic models for human, mouse, rat, dog, or rabbit.
}
\usage{tissue.data}
\format{A data.frame containing 13 rows and 20 columns.}
\source{Wambaugh et al. (2014), in preparation}
\references{
Birnbaum, L and Brown, R and Bischoff, K and Foran, J and Blancato, J and Clewell, H and Dedrick, R (1994). Physiological parameter values for PBPK model. International Life Sciences Institute, Risk Science Institute, Washington, DC


Schmitt, W. (2008). General approach for the calculation of tissue to plasma partition coefficients. Toxicology in vitro : an international journal published in association with BIBRA 22(2), 457-67, 10.1016/j.tiv.2007.09.010.



}
\author{
John Wambaugh and Nisha Sipes
}
\keyword{ datasets }
