% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_scale.R
\name{tissue_scale}
\alias{tissue_scale}
\title{Allometric scaling.}
\usage{
tissue_scale(height_ref, height_indiv, tissue_mean_ref)
}
\arguments{
\item{height_ref}{Reference height in cm.}

\item{height_indiv}{Individual height in cm.}

\item{tissue_mean_ref}{Reference tissue mass or flow.}
}
\value{
Allometrically scaled tissue mass or flow, in the same units as
 \code{tissue_mean_ref}.
}
\description{
Allometrically scale a tissue mass or flow based on height^{3/4}.
}
\references{
\insertRef{ring2017identifying}{httk}
}
\author{
Caroline Ring
}
\keyword{httk-pop}
