% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truong25.seem3}
\alias{truong25.seem3}
\title{SEEM3 Example Data for Truong et al. 2025}
\format{
data.table and data.frame
}
\usage{
truong25.seem3
}
\description{
We can grab SEEM daily intake rate predictions already in RData format from
https://github.com/HumanExposure/SEEM3RPackage/blob/main/scripts/
Download the file chem.preds-2018-11-28.RData
}
\details{
We do not have the space to distribute all the SEEM predictions within
this R package, but we can give you our "Full Human Gestational IVIVE" example chemicals.
}
\references{
\insertRef{truong2025fullpregnancy}{httk}

\insertRef{Ring2017SEEM}{httk}
}
\keyword{data}
