% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.nhanes}
\alias{wambaugh2019.nhanes}
\title{NHANES Chemical Intake Rates for chemicals in Wambaugh et al. (2019)}
\format{A data frame with 20 rows and 4 variables:
\describe{
  \item{lP}{The median of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{lP.min}{The lower 95th percentile of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{lP.max}{The upper 95th percentile of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{CASRN}{The Chemical Abstracts Service Registry Number}
}}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.nhanes
}
\description{
These data are a subset of the Bayesian inferrences reported by Ring et al.
(2017) from the U.S. Centers for Disease Control and Prevention (CDC)
National Health and Nutrition Examination Survey (NHANES). They reflect the
populaton median intake rate (mg/kg body weight/day), with uncertainty.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
evironmental chemicals by simulating toxicokinetic variability." Environment
international 106 (2017): 105-118

Wambaugh et al. (2019) "Assessing Toxicokinetic Uncertainty and
Variability in Risk Prioritization", Toxicological Sciences, 172(2), 235-251.
}
\author{
John Wambaugh
}
\keyword{data}
