% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.seem3}
\alias{wambaugh2019.seem3}
\title{ExpoCast SEEM3 Consensus Exposure Model Predictions for Chemical Intake Rates}
\format{
A data frame with 385 rows and 38 variables:
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.seem3
}
\description{
These data are a subset of the Bayesian inferrences reported by Ring et al.
(2019) for a consensus model of twelve exposue predictors. The predictors
were calibrated based upon their ability to predict intake rates inferred
National Health and Nutrition Examination Survey (NHANES). They reflect the
populaton median intake rate (mg/kg body weight/day), with uncertainty.
}
\references{
\insertRef{Ring2017SEEM}{httk}

\insertRef{wambaugh2019assessing}{httk}
}
\author{
John Wambaugh
}
\keyword{data}
