% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wfl}
\alias{wfl}
\title{WHO weight-for-length charts}
\format{
a data.table with 262 rows and 4 variables:

 \describe{ \item{Sex}{"Male" or "Female"} \item{Length}{Recumbent length in
 cm} \item{P2.3}{The 2.3rd percentile weight in kg for the corresponding sex
 and recumbent length} \item{P97.7}{The 97.7th percentile weight in kg for
 the corresponding sex and recumbent length}}
}
\source{
\url{https://www.cdc.gov/growthcharts/who/boys_weight_head_circumference.htm}
and
\url{https://www.cdc.gov/growthcharts/who/girls_weight_head_circumference.htm}
}
\usage{
wfl
}
\description{
Charts giving weight-for-length percentiles for boys and girls under age 2.
}
\details{
For infants under age 2, weight class depends on weight for length percentile.
#'\describe{ \item{Underweight}{<2.3rd percentile} \item{Normal
weight}{2.3rd-97.7th percentile} \item{Obese}{>=97.7th percentile} }
}
\keyword{datasets}
