% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fup_adjustment.R
\name{apply_fup_adjustment}
\alias{apply_fup_adjustment}
\title{Correct the measured intrinsive hepatic clearance for fraction free}
\usage{
apply_fup_adjustment(
  fup,
  fup.correction = NULL,
  Pow = NULL,
  pKa_Donor = NULL,
  pKa_Accept = NULL,
  suppress.messages = FALSE,
  minimum.Funbound.plasma = 1e-04
)
}
\arguments{
\item{Pow}{The octanal:water equilibrium partition coefficient}

\item{pKa_Donor}{A string containing hydrogen donor ionization equilibria, 
concatenated with commas. Can be "NA" if none exist.}

\item{pKa_Accept}{A string containing hydrogen acceptance ionization equilibria, 
concatenated with commas. Can be "NA" if none exist.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{minimum.Funbound.plasma}{\eqn{f_{up}} is not allowed to drop below
this value (default is 0.0001).}

\item{Clint}{In vitro measured intrinsic hepatic clearance in units of
(ul/min/million hepatocytes).

#param Fu_hep Estimated fraction of chemical free for metabolism in the 
in vitro assay, estimated by default from the method of Kilford et al. (2008)
using \code{\link{calc_hep_fu}}}
}
\value{
Fraction unbound in plasma adjusted to take into account binding
in the in vitro assay
}
\description{
This function uses the free fraction estimated from Kilford et al. (2008) 
to increase the in vitro measure intrinsic hepatic clearance. The assumption
that chemical that is bound in vitro is not available to be metabolized and
therefore the actual rate of clearance is actually faster. Note that in most
high throughput TK models included in the package this increase is offset by
the assumption of "restrictive clearance" -- that is, the rate of hepatic
metabolism is slowed to account for the free fraction of chemical in plasma.
This adjustment was made starting in Wetmore et al. (2015) in order to better
predict plasma concentrations.
}
\references{
Kilford, Peter J., et al. "Hepatocellular binding of drugs: 
correction for unbound fraction in hepatocyte incubations using microsomal 
binding or drug lipophilicity data." Drug Metabolism and Disposition 36.7 
(2008): 1194-1197.

Wetmore, Barbara A., et al. "Incorporating high-throughput exposure 
predictions with dosimetry-adjusted in vitro bioactivity to inform chemical 
toxicity testing." Toxicological Sciences 148.1 (2015): 121-136.
}
\seealso{
\code{\link{calc_help_fu}}
}
\author{
John Wambaugh
}
\keyword{in-vitro}
