% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armitage.R
\name{armitage_estimate_sarea}
\alias{armitage_estimate_sarea}
\title{Estimate well surface area}
\usage{
armitage_estimate_sarea(
  tcdata = NA,
  this.well_number = 384,
  this.cell_yield = NA,
  this.v_working = NA
)
}
\arguments{
\item{tcdata}{A data table with well_number corresponding to plate format,
optionally include v_working, sarea, option.bottom, and option.plastic}

\item{this.well_number}{For single value, plate format default is 384, used
if is.na(tcdata)==T}

\item{this.cell_yield}{For single value, optionally supply cell_yield,
otherwise estimated based on well number}

\item{this.v_working}{For single value, optionally supply working volume,
otherwise estimated based on well number (m^3)}
}
\value{
tcdata, A data table with well_number, sarea (surface area, m^2),
cell_yield (# cells), v_working (m^3), v_total (m^3) per well
}
\description{
Estimate geometry surface area of plastic in well plate based on well plate
format suggested values from Corning.  option.plastic == T (default) give
nonzero surface area (sarea, m^2) option.bottom == T (default) includes
surface area of the bottom of the well in determining sarea.  Optionally
include user values for working volume (v_working, m^3) and surface area.
}
\author{
Greg Honda
}
