% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_httk.R
\name{benchmark_httk}
\alias{benchmark_httk}
\title{Assess the curent performance of httk relative to historical benchmarks}
\usage{
benchmark_httk(
  basic.check = TRUE,
  calc_mc_css.check = TRUE,
  in_vivo_css.check = TRUE,
  suppress.messages = TRUE,
  make.plots = TRUE
)
}
\arguments{
\item{basic.check}{Whether to run the basic checks, including units uM and 
mg/L units for calc_analytic_css, calc_mc_css, and solve_pbtk as well as 
the number of chemicals with sufficient data to run the steady_stat emodel 
(defaults to TRUE)}

\item{calc_mc_css.check}{A character string of three numbers separated by two 
dashes (defaults to TRUE)}

\item{in_vivo_css.check}{A character string of three numbers separated by two 
dashes (defaults to TRUE)}

\item{suppress.messages}{Whether or not output messages are suppressed
(defaults to TRUE)}

\item{make.plots}{Whether current benchmarks should be plotted with 
historical performance (defaults to TRUE)}
}
\value{
named list, whose elements depend on the selected checks
\tabular{rl}{
  basic \tab A list with four metrics:
  \itemize{
    \item{N.steadystate}{Number of chemicals with sufficient data for steady-state IVIVE}
    \item{calc_analytic.units}{Ratio of mg/L to uM * 1000 / molecular weight -- should be 1}
    \item{calc_mc.units}{Ratio should be 1}
    \item{solve_pbtk.units}{Ratio should be 1}
  } \cr
  calc_mc_css \tab A list with four metrics:
  \itemize{
    \item{RMSLE.Wetmore}{Root mean squared log10 error (RMSLE) in predicted Css between literature valuse (SimCyp, Wetmore et al. 2012,2015) and calc_mc_css}
    \item{N.Wetmore}{Number of chemicals in Wetmore comparison}
    \item{RMSLE.noMC}{RMSLE between calc_analytic_css and calc_mc_css}
    \item{N.noMC}{Number of chemicals in noMC comparison}
  } \cr
  in_vivo_css \tab A list with two metrics:
  \itemize{
    \item{RMSLE.InVivoCss}{RMSLE between the predictions of calc_analytic_css and in vivo estimates of Css}
    \item{N.InVivoCss}{Number of chemcials in comparison}
  } \cr                    
}
}
\description{
The function performs a series of "sanity checks" and predictive performance
benchmarks so that the impace of changes to the data, models, and 
implementation of the R package can be tested. Plots can be generated showing 
how the performance of he current version compares with past releases of
httk.
}
\details{
Historically some refinements made to one aspect of httk have unitentiionally
impacted other aspects. Most notably errors have ocasionally been introduced 
with respect to units (v1.9, v2.1.0). This benchmarking tool is intended to
reduce the chance of these errors occuring in the future.
}
\references{
Wambaugh et al. "Developing Generic Toxicokinetic Models with R Package 
"httk" for Enhanced Reporting Accuracy and Statistical Evaluation",
in preparation
}
\author{
John Wambaugh
}
