% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_analytic_css_sumclearances.R
\name{calc_analytic_css_sumclearances}
\alias{calc_analytic_css_sumclearances}
\title{Calculate the steady state concentration for the sum of clearances
steady-state model with exhalation}
\usage{
calc_analytic_css_sumclearances(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  parameters = NULL,
  dosing = list(daily.dose = 1),
  hourly.dose = NULL,
  dose.units = "mg",
  concentration = "plasma",
  species = "human",
  Caco2.options = NULL,
  suppress.messages = FALSE,
  recalc.blood2plasma = FALSE,
  tissue = NULL,
  route = "oral",
  restrictive.clearance = TRUE,
  bioactive.free.invivo = FALSE,
  ...
)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or the parameters must 
be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must 
be specified.}

\item{dtxsid}{EPA's 'DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{parameters}{Chemical parameters from \code{\link{parameterize_sumclearances}} overrides chem.name and chem.cas.}

\item{dosing}{List of dosing metrics used in simulation, which includes
the namesake entries of a model's associated dosing.params. For steady-state
calculations this is likely to be either "daily.dose" for oral exposures or
"Cinhaled" for inhalation.}

\item{hourly.dose}{Hourly dose rate mg/kg BW/h.}

\item{dose.units}{The units associated with the dose received.}

\item{concentration}{Desired concentration type, 'blood' or default 'plasma'.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{Caco2.options}{A list of options to use when working with Caco2 apical to
basolateral data \code{Caco2.Pab}, default is Caco2.options = list(Caco2.Pab.default = 1.6,
Caco2.Fabs = TRUE, Caco2.Fgut = TRUE, overwrite.invivo = FALSE, keepit100 = FALSE). Caco2.Pab.default sets the default value for 
Caco2.Pab if Caco2.Pab is unavailable. Caco2.Fabs = TRUE uses Caco2.Pab to calculate
fabs.oral, otherwise fabs.oral = \code{Fabs}. Caco2.Fgut = TRUE uses Caco2.Pab to calculate 
fgut.oral, otherwise fgut.oral = \code{Fgut}. overwrite.invivo = TRUE overwrites Fabs and Fgut in vivo values from literature with 
Caco2 derived values if available. keepit100 = TRUE overwrites Fabs and Fgut with 1 (i.e. 100 percent) regardless of other settings.
See \code{\link{get_fbio}} for further details.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{recalc.blood2plasma}{Recalculates the ratio of the amount of chemical 
in the blood to plasma using the input parameters. Use this if you have 
'altered hematocrit, Funbound.plasma, or Krbc2pu.}

\item{tissue}{Desired tissue concentration (defaults to whole body 
concentration.)}

\item{route}{Route of exposure ("inhalation" or [DEFAULT] "oral").}

\item{restrictive.clearance}{If TRUE (default), then only the fraction of
chemical not bound to protein is available for metabolism in the liver. If 
FALSE, then all chemical in the liver is metabolized (faster metabolism due
to rapid off-binding).}

\item{bioactive.free.invivo}{If FALSE (default), then the total concentration is treated
as bioactive in vivo. If TRUE, the the unbound (free) plasma concentration is treated as 
bioactive in vivo. Only works with tissue = NULL in current implementation.}

\item{...}{Additional parameters passed to parameterize function if 
parameters is NULL.}
}
\value{
Steady state plasma concentration in mg/L units
}
\description{
This function calculates the analytic steady state plasma or venous blood 
concentrations as a result of infusion dosing.
}
\seealso{
\code{\link{calc_analytic_css}}

\code{\link{parameterize_steadystate}}
}
\author{
John Wambaugh
}
\keyword{sumclearances}
