% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionization_functions.R
\name{calc_dow}
\alias{calc_dow}
\title{Calculate the distribution coefficient}
\usage{
calc_dow(
  Pow = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  parameters = NULL,
  pH = NULL,
  pKa_Donor = NULL,
  pKa_Accept = NULL,
  fraction_charged = NULL,
  alpha = 0.001
)
}
\arguments{
\item{Pow}{Octanol:water partition coefficient (ratio of concentrations)}

\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{dtxsid}{EPA's 'DSSTox Structure ID (https://comptox.epa.gov/dashboard)
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{parameters}{Chemical parameters from a parameterize_MODEL function,
overrides chem.name and chem.cas.}

\item{pH}{pH where ionization is evaluated.}

\item{pKa_Donor}{Compound H dissociation equilibirum constant(s).
Overwrites chem.name and chem.cas.}

\item{pKa_Accept}{Compound H association equilibirum constant(s).
Overwrites chem.name and chem.cas.}

\item{fraction_charged}{Fraction of chemical charged at the given pH}

\item{alpha}{Ratio of Distribution coefficient D of totally charged species and that of the neutral form}
}
\value{
Distribution coefficient (numeric)
}
\description{
This function estimates the ratio of the equilibrium concentrations of
a compound in octanol and water, taking into account the charge of the
compound. Given the pH, we assume the neutral (uncharged) fraction of
compound partitions according to the hydrophobicity (Pow). We assume that
only a fraction alpha (defaults to 0.001 -- Schmitt (2008)) of the charged
compound partitions into lipid (octanol). Fractions charged are calculated
according to hydrogen ionization equilibria (pKa_Donor, pKa_Accept) using
\code{\link{calc_ionization}}.
}
\references{
Schmitt, Walter. "General approach for the calculation of tissue to plasma 
partition coefficients." Toxicology in vitro 22.2 (2008): 457-467.

Pearce, Robert G., et al. "Evaluation and calibration of
high-throughput predictions of chemical distribution to tissues." Journal of
Pharmacokinetics and Pharmacodynamics 44.6 (2017): 549-565.

Strope, Cory L., et al. "High-throughput in-silico prediction of ionization 
equilibria for pharmacokinetic modeling." Science of The Total Environment 
615 (2018): 150-160.
}
\seealso{
\link{\code{calc_ionization}}
}
\author{
Robert Pearce and John Wambaugh
}
\keyword{Parameter}
