% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_oral_bioavailability.R
\name{calc_fbio.oral}
\alias{calc_fbio.oral}
\alias{calc_fabs.oral}
\alias{calc_fgut.oral}
\title{Caco2 Oral Fraction Calculators}
\usage{
calc_fbio.oral(
  Params = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  suppress.messages = FALSE,
  Caco2.Pab.default = "1.6",
  Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE,
  overwrite.invivo = FALSE,
  keepit100 = FALSE,
  ...
)

calc_fabs.oral(
  Params = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  suppress.messages = FALSE,
  Caco2.Pab.default = "1.6",
  Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE,
  overwrite.invivo = FALSE,
  keepit100 = FALSE,
  ...
)

calc_fgut.oral(
  Params = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  suppress.messages = FALSE,
  Caco2.Pab.default = "1.6",
  Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE,
  overwrite.invivo = FALSE,
  keepit100 = FALSE,
  ...
)
}
\arguments{
\item{Params}{A list of the parameters (Caco2.Pab, Funbound.Plasma, Rblood2plasma,
Clint, BW, Qsmallintestine, Fabs, Fgut) used in the calculation, either supplied by user
or calculated in parameterize_steady_state.}

\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{default.to.human}{Substitutes missing rat values with human values if
true.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{Caco2.Pab.default}{= "1.6" Caco2 apical to basolateral data}

\item{Caco2.Fgut}{= TRUE uses Caco2.Pab to calculate fgut.oral, otherwise fgut.oral = \code{Fgut}}

\item{Caco2.Fabs}{= TRUE uses Caco2.Pab to calculate fabs.oral, otherwise fabs.oral = \code{Fabs}.}

\item{overwrite.invivo}{= TRUE overwrites Fabs and Fgut in vivo values from literature with}

\item{keepit100}{= TRUE overwrites Fabs and Fgut with 1 (i.e. 100 percent) regardless of other settings.

Caco2 derived values if available.}

\item{...}{Other parameters}
}
\value{
\item{fbio.oral}{Oral bioavailability, the fraction of oral dose 
reaching systemic distribution in the body.} \item{fabs.oral}{Fraction of dose absorbed, 
i.e. the fraction of the dose that enters the gutlumen.} \item{fgut.oral}{Fraction of 
chemical surviving first pass metabolism in the gut.} \item{fhep.oral}{Fraction of chemical
surviving first pass hepatic clearance.}
}
\description{
Caco2 is related to effective permeability based on Yang et al. (2007)
These functions calculate the fraction absorbed (calc_fabs.oral -- 
Darwich et al. (2010)), the fraction
surviving first pass gut metabolism (calc_fgut.oral), and the oral bioavailability
(calc_fbio.oral). Note that the first pass hepatic clearance is calculated within the
parameterization and other functions. using \code{\link{calc_hep_bioavailability}}
}
\section{Functions}{
\itemize{
\item \code{calc_fabs.oral()}: Calculate the fraction absorbed in the gut (Darwich et al., 2010)

\item \code{calc_fgut.oral()}: Calculate the fraction of chemical surviving first pass metabolism in the gut

}}
\examples{

fbio1 <- calc_fbio.oral(chem.cas='80-05-7',
                        Caco2.Pab.default = 2,
                        Caco2.Fabs = TRUE,
                        Caco2.Fgut = TRUE)
fbio2 <- calc_fbio.oral(chem.cas='80-05-7',
                        Caco2.Pab.default = 2,
                        Caco2.Fabs = FALSE,
                        Caco2.Fgut = FALSE)

}
\references{
\insertRef{darwich2010interplay}{httk}
\insertRef{yang2007prediction}{httk}
\insertRef{HondaUnpublishedCaco2}{httk}
}
\author{
Gregory Honda
}
\keyword{Parameter}
