% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_correction.R
\name{calc_fup_correction}
\alias{calc_fup_correction}
\title{Calculate the correction for lipid binding in plasma binding assay}
\usage{
calc_fup_correction(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  parameters = NULL,
  Flipid = NULL,
  plasma.pH = 7.4
)
}
\arguments{
\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{dtxsid}{EPA's 'DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
 -- if parameters is not specified then the chemical must be identified by 
either CAS, name, or DTXSIDs}

\item{parameters}{Parameters from the appropriate parameterization function
for the model indicated by argument model}

\item{Vr}{Rratio of cell volume to incubation volume. Default is taken from}

\item{pH}{pH of the incupation medium.}
}
\value{
A numeric fraction unpbound in plasma between zero and one
}
\description{
Pearce et al. (2017) compared predicted and observed partition coefficients
for a range of compounds. Based on observed discrepancies they 
argued that there should be greater lipid binding in plasma (in vivo) 
for some lipophilic un-ionized compounds than is observed in vitro.
Following Schmitt (2008), the lipid binding in tissue is assumed to follow
the distribution coefficient 
\deqn{D = K_{nL} = P_{ow}*(F_{neutral} + \alpha*F_{charged})}
and the fractional volume of lipid in
plasma. The error in \eqn{K_{tissue:plasma}} prediction for these compounds
has been hypothesized to be lower due to insufficient lipid
binding in the in vitro protein binding assay (Poulin and Haddad (2012)).
This deficiency in the assay leads to a much higher
unbound concentration relative to the bound concentration
for lipophilic compounds and thus an overestimation of \eqn{f_up}
and \eqn{K_{tissue:plasma}}. To adjust for this we recalculated the plasma:water
partition coefficient (\eqn{1/f_{up}}) to account for neutral lipid
binding in plasma, assuming that some lipid is absent from
the in vitro assay:
\deqn{f^{corrected}_{up} = \frac{1}{K_{nL}^{pl}*F_{lipid} + \frac{1}{f^{in vitro}_{up}}}}
}
\references{
Pearce, Robert G., et al. "Evaluation and calibration of high-throughput 
predictions of chemical distribution to tissues." Journal of pharmacokinetics 
and pharmacodynamics 44.6 (2017): 549-565.

Poulin, Patrick, and Sami Haddad. "Advancing prediction of tissue 
distribution and volume of distribution of highly lipophilic compounds from 
a simplified tissue-composition-based model as a mechanistic animal 
alternative method." Journal of pharmaceutical sciences 101.6 (2012): 
2250-2261.

Schmitt, Walter. "General approach for the calculation of 
tissue to plasma partition coefficients." Toxicology in Vitro 22.2 (2008): 
457-467.
}
\seealso{
\code{\link{adjust_fup}}
}
\author{
John Wambaugh
}
\keyword{in-vitro}
