% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_maternal_bw.R
\name{calc_maternal_bw}
\alias{calc_maternal_bw}
\title{Calculate maternal body weight}
\usage{
calc_maternal_bw(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  week = 12,
  ...
)
}
\arguments{
\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{...}{Arguments passed to parameterize_pbtk.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human"). Currently only a narrow human model is supported.}
}
\value{
\item{BW}{Maternal Body Weight, kg.}
}
\description{
This function initializes the parameters needed in the functions
solve_fetal_pbtk by calling solve_pbtk and adding additional parameters.
}
\details{
BW <- params$pre_pregnant_BW + 
   params$BW_cubic_theta1 * tw + 
   params$BW_cubic_theta2 * tw^2 + 
   params$BW_cubic_theta3 * tw^3
}
\examples{

}
\references{
Kapraun, Dustin F., et al. "Empirical models for anatomical and physiological 
changes in a human mother and fetus during pregnancy and gestation." 
PloS one 14.5 (2019): e0215906.
}
\author{
John Wambaugh
}
\keyword{Parameter}
