% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{convert_units}
\usage{
convert_units(
  input.units = NULL,
  output.units = NULL,
  MW = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  parameters = NULL
)
}
\arguments{
\item{input.units}{Assigned input units of interest}

\item{output.units}{Desired output units}

\item{MW}{Molecular weight of substance of interest in g/mole}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{dtxsid}{EPA's DSSTox Structure ID 
(\url{http://comptox.epa.gov/dashboard}) the chemical must be identified by
either CAS, name, or DTXSIDs}

\item{parameters}{A set of model parameters, especially a set that
includes MW (molecular weight) for our conversions}
}
\description{
This function is designed to accept input units, output units, and the 
molecular weight (MW) of a substance of interest to then use a table lookup
to return a scaling factor that can be readily applied for the intended
conversion. It can also take chemical identifiers in the place of a 
specified molecular weight value to retrieve that value for its own use.
}
\details{
If input or output units not contained in the table are queried,
it gives a corresponding error message. It gives a warning message about the
handling of 'ppmv,' as the function is only set up to convert between ppmv 
and mass-based units (like mg/m^3 or umol/L) in the context of ideal gases.

convert_units is not directly configured to accept and convert units based
on BW, like mg/kg. For this purpose, see \code{\link{scale_dosing}}.

The function supports a limited set of most relevant units across
toxicological models, currently including umol, uM, mg, mg/L, mg/m^3, and
in the context of gases assumed to be ideal, ppmv.
}
\examples{

MW_pyrene <- get_physchem_param(param <- 'MW', chem.name <- 'pyrene')
conversion_factor <- convert_units(input.units <- 'mg/L', output.units <- 'uM',
MW <- MW_pyrene)

}
\author{
Mark Sfeir and John Wambaugh
}
