% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gfr.R
\name{estimate_gfr}
\alias{estimate_gfr}
\title{Predict GFR.}
\usage{
estimate_gfr(gfrtmp.dt)
}
\arguments{
\item{gfrtmp.dt}{A data.table with columns \code{gender}, \code{reth}, 
\code{age_years}, \code{age_months}, \code{BSA_adj}, \code{serum_creat}.}
}
\value{
The same data.table with a \code{gfr_est} column added, containing 
 estimated GFR values.
}
\description{
First predict serum creatinine using smoothing spline, then predict GFR using
CKD-EPI equation.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to 
environmental chemicals by simulating toxicokinetic variability." Environment 
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{httk-pop}
