% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cheminfo.R
\name{get_cheminfo}
\alias{get_cheminfo}
\title{Retrieve chemical information from HTTK package}
\usage{
get_cheminfo(
  info = "CAS",
  species = "Human",
  fup.lod.default = 0.005,
  model = "3compartmentss",
  default.to.human = FALSE,
  median.only = FALSE,
  fup.ci.cutoff = TRUE,
  clint.pvalue.threshold = 0.05,
  suppress.messages = FALSE
)
}
\arguments{
\item{info}{A single character vector (or collection of character vectors)
from "Compound", "CAS", "DTXSID, "logP", "pKa_Donor"," pKa_Accept", "MW", "Clint",
"Clint.pValue", "Funbound.plasma","Structure_Formula", or "Substance_Type". info="all"
gives all information for the model and species.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{fup.lod.default}{Default value used for fraction of unbound plasma for
chemicals where measured value was below the limit of detection. Default
value is 0.0005.}

\item{model}{Model used in calculation, 'pbtk' for the multiple compartment
model, '1compartment' for the one compartment model, '3compartment' for
three compartment model, '3compartmentss' for the three compartment model
without partition coefficients, or 'schmitt' for chemicals with logP and
fraction unbound (used in predict_partitioning_schmitt).}

\item{default.to.human}{Substitutes missing values with human values if
true.}

\item{median.only}{Use median values only for fup and clint.  Default is FALSE.}

\item{fup.ci.cutoff}{Cutoff for the level of uncertainty in fup estimates.
This value should be between (0,1). Default is `NULL` specifying no filtering.}

\item{clint.pvalue.threshold}{Hepatic clearance for chemicals where the in
vitro clearance assay result has a p-values greater than the threshold are
set to zero.}

\item{suppress.messages}{Whether or not the output messages are suppressed.}
}
\value{
\item{info}{Table/vector containing values specified in "info" for
valid chemicals.}
}
\description{
This function provides the information specified in "info=" (can be single entry
or vector) for all chemicals for which a toxicokinetic model can be
paramterized for a given species.
}
\details{
When default.to.human is set to TRUE, and the species-specific data,
Funbound.plasma and Clint, are missing from chem.physical_and_invitro.data,
human values are given instead.
}
\examples{

\donttest{
# List all CAS numbers for which the 3compartmentss model can be run in humans: 
get_cheminfo()

get_cheminfo(info=c('compound','funbound.plasma','logP'),model='pbtk') 
# See all the data for humans:
get_cheminfo(info="all")

TPO.cas <- c("741-58-2", "333-41-5", "51707-55-2", "30560-19-1", "5598-13-0", 
"35575-96-3", "142459-58-3", "1634-78-2", "161326-34-7", "133-07-3", "533-74-4", 
"101-05-3", "330-54-1", "6153-64-6", "15299-99-7", "87-90-1", "42509-80-8", 
"10265-92-6", "122-14-5", "12427-38-2", "83-79-4", "55-38-9", "2310-17-0", 
"5234-68-4", "330-55-2", "3337-71-1", "6923-22-4", "23564-05-8", "101-02-0", 
"140-56-7", "120-71-8", "120-12-7", "123-31-9", "91-53-2", "131807-57-3", 
"68157-60-8", "5598-15-2", "115-32-2", "298-00-0", "60-51-5", "23031-36-9", 
"137-26-8", "96-45-7", "16672-87-0", "709-98-8", "149877-41-8", "145701-21-9", 
"7786-34-7", "54593-83-8", "23422-53-9", "56-38-2", "41198-08-7", "50-65-7", 
"28434-00-6", "56-72-4", "62-73-7", "6317-18-6", "96182-53-5", "87-86-5", 
"101-54-2", "121-69-7", "532-27-4", "91-59-8", "105-67-9", "90-04-0", 
"134-20-3", "599-64-4", "148-24-3", "2416-94-6", "121-79-9", "527-60-6", 
"99-97-8", "131-55-5", "105-87-3", "136-77-6", "1401-55-4", "1948-33-0", 
"121-00-6", "92-84-2", "140-66-9", "99-71-8", "150-13-0", "80-46-6", "120-95-6",
"128-39-2", "2687-25-4", "732-11-6", "5392-40-5", "80-05-7", "135158-54-2", 
"29232-93-7", "6734-80-1", "98-54-4", "97-53-0", "96-76-4", "118-71-8", 
"2451-62-9", "150-68-5", "732-26-3", "99-59-2", "59-30-3", "3811-73-2", 
"101-61-1", "4180-23-8", "101-80-4", "86-50-0", "2687-96-9", "108-46-3", 
"95-54-5", "101-77-9", "95-80-7", "420-04-2", "60-54-8", "375-95-1", "120-80-9",
"149-30-4", "135-19-3", "88-58-4", "84-16-2", "6381-77-7", "1478-61-1", 
"96-70-8", "128-04-1", "25956-17-6", "92-52-4", "1987-50-4", "563-12-2", 
"298-02-2", "79902-63-9", "27955-94-8")
httk.TPO.rat.table <- subset(get_cheminfo(info="all",species="rat"),
 CAS \%in\% TPO.cas)
 
httk.TPO.human.table <- subset(get_cheminfo(info="all",species="human"),
 CAS \%in\% TPO.cas)
}

}
\author{
John Wambaugh, Robert Pearce, and Sarah E. Davidson
}
\keyword{Retrieval}
