% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fabsgut.R
\name{get_fabsgut}
\alias{get_fabsgut}
\title{get_fabsgut}
\usage{
get_fabsgut(
  Params = NULL,
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  Caco2.Pab.default = "1.6",
  Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE,
  overwrite.invivo = FALSE,
  keepit100 = FALSE
)
}
\arguments{
\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- the 
chemical must be identified by either CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  the 
chemical must be identified by either CAS, name, or DTXISD}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})
-- the chemical must be identified by either CAS, name, or DTXSIDs}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{default.to.human}{Substitutes missing rat values with human values if
true.}

\item{Caco2.Fgut}{= TRUE uses Caco2.Pab to calculate  fgut.oral, otherwise 
fgut.oral = \code{Fgut}.}

\item{Caco2.Fabs}{= TRUE uses Caco2.Pab to calculate
fabs.oral, otherwise fabs.oral = \code{Fabs}.}

\item{overwrite.invivo}{= TRUE overwrites Fabs and Fgut in vivo values from 
literature with Caco2 derived values if available.}

\item{keepit100}{TRUE overwrites Fabs and Fgut with 1 (i.e. 100 percent) 
regardless of other settings.}

\item{Caco2.default}{sets the default value for  Caco2.Pab if Caco2.Pab is 
unavailable.}
}
\description{
This function uses Caco2 permeability to estimate oral absorption fraction.
}
\author{
Greg Honda and John Wambaugh
}
\keyword{Parameter}
\keyword{oral_bioavailability}
