% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invitroPK_param.R
\name{get_invitroPK_param}
\alias{get_invitroPK_param}
\title{Retrieve species-specific in vitro data from chem.physical_and_invitro.data table}
\usage{
get_invitroPK_param(
  param,
  species,
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL
)
}
\arguments{
\item{param}{The in vitro pharmacokinetic parameter needed.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}
}
\value{
The value of the parameter, if found
}
\description{
This function retrieves in vitro PK data (for example intrinsic metabolic clearance 
or fraction unbound in plasma) from the main HTTK data. This function looks
for species-specific values.
}
\details{
Note that this function works with a local version of the 
get.physical_and_invitro.data table to allow users to add/modify chemical
data (for example, via \code{\link{add_chemtabke}} or 
\code{\link{load_sipes2017}}).
}
\seealso{
\code{\link{get_physchem_param}}
}
\author{
John Wambaugh and Robert Pearce
}
