% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk-package.R
\docType{package}
\name{httk-package}
\alias{httk}
\alias{httk-package}
\title{httk: High-Throughput Toxicokinetics}
\description{
This is a modified version of the EPA's 'httk' R package (v2.7.3) (github.com/USEPA/CompTox-ExpoCast-httk). This version of 'httk' removes the use of global variables within the package. A new function argument 'chemdata' was added to many functions and its default value was set to httk's 'chem.physical_and_invitro.data' data frame. If a user needs to add new chemicals to simulate, a new data frame (original chem.physical_and_invitro.data plus rows with new chemicals) can be just passed through the function instead of using the global environment to store it. The remaining variables originally defined as 'global' have been made available as internal data sets. No algorithms have been changed from the original 'httk' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KristenWindoloski/CompTox-ExpoCast-httk}
}

}
\author{
\strong{Maintainer}: John Wambaugh \email{wambaugh.research@gmail.com} (\href{https://orcid.org/0000-0002-4024-534X}{ORCID})

Authors:
\itemize{
  \item Sarah Davidson-Fritz \email{davidsonfritz.sarah@epa.gov} (\href{https://orcid.org/0000-0002-2891-9380}{ORCID})
  \item Robert Pearce (\href{https://orcid.org/0000-0003-3168-4049}{ORCID})
  \item Caroline Ring \email{Ring.Caroline@epa.gov} (\href{https://orcid.org/0000-0002-0463-1251}{ORCID})
  \item Greg Honda \email{honda.gregory@epa.gov} (\href{https://orcid.org/0000-0001-7713-9850}{ORCID})
  \item Mark Sfeir
  \item Matt Linakis \email{MLINAKIS@ramboll.com} (\href{https://orcid.org/0000-0003-0526-2395}{ORCID})
  \item Dustin Kapraun \email{kapraun.dustin@epa.gov} (\href{https://orcid.org/0000-0001-5570-6383}{ORCID})
  \item Kimberly Truong \email{truong.kimberly@epa.gov} (\href{https://orcid.org/0000-0003-1716-1006}{ORCID})
  \item Colin Thomson \email{thomson.colin@epa.gov} (\href{https://orcid.org/0000-0001-5303-6646}{ORCID})
  \item Meredith Scherer \email{Scherer.Meredith@epa.gov} (\href{https://orcid.org/0000-0002-9785-7363}{ORCID})
  \item Annabel Meade \email{aemeade7@gmail.com} (\href{https://orcid.org/0000-0002-5452-8797}{ORCID})
  \item Celia Schacht \email{Schacht.Celia@epa.gov} (\href{https://orcid.org/0000-0002-8131-513X}{ORCID})
  \item Leonie Lautz \email{leonie.lautz@wur.nl} (\href{https://orcid.org/0000-0003-4648-666X}{ORCID})
}

Other contributors:
\itemize{
  \item Kristen Windoloski \email{kawindoloski@gmail.com} [contributor]
  \item Todor Antonijevic \email{tantonijevic@toxstrategies.com} (\href{https://orcid.org/0000-0002-0248-8412}{ORCID}) [contributor]
  \item Miyuki Breen \email{breen.miyuki@epa.gov} (\href{https://orcid.org/0000-0001-8511-4653}{ORCID}) [contributor]
  \item Shannon Bell \email{Shannon.bell@inotivco.com} (\href{https://orcid.org/0000-0002-5754-6085}{ORCID}) [contributor]
  \item Xiaoqing Chang \email{Xiaoqing.chang@inotivco.com} (\href{https://orcid.org/0000-0003-0752-1848}{ORCID}) [contributor]
  \item Jimena Davis [contributor]
  \item Elaina Kenyon \email{kenyon.elaina@epa.gov} (\href{https://orcid.org/0000-0001-8781-2391}{ORCID}) [contributor]
  \item Gilberto Padilla Mercado \email{PadillaMercado.Gilberto@epa.gov} (\href{https://orcid.org/0000-0001-5423-1646}{ORCID}) [contributor]
  \item Katie Paul Friedman \email{Paul-Friedman.Katie@epa.gov} (\href{https://orcid.org/0000-0002-2710-1691}{ORCID}) [contributor]
  \item Nathan Pollesch \email{pollesch.nathan@epa.gov} (\href{https://orcid.org/0000-0001-8992-9294}{ORCID}) [contributor]
  \item Noelle Sinski \email{Noelle.Sinski@icf.com} [contributor]
  \item Nisha Sipes \email{sipes.nisha@epa.gov} (\href{https://orcid.org/0000-0003-4203-6426}{ORCID}) [contributor]
  \item James Sluka \email{jsluka@indiana.edu} (\href{https://orcid.org/0000-0002-5901-1404}{ORCID}) [contributor]
  \item Caroline Stevens \email{stevens.caroline@epa.gov} (\href{https://orcid.org/0000-0001-7780-2691}{ORCID}) [contributor]
  \item Barbara Wetmore \email{wetmore.barbara@epa.gov} (\href{https://orcid.org/0000-0002-6878-5348}{ORCID}) [contributor]
  \item Lily Whipple [contributor]
  \item Woodrow Setzer (\href{https://orcid.org/0000-0002-6709-9186}{ORCID}) [contributor]
}

}
\keyword{internal}
