% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk-package.R
\docType{package}
\name{httk-package}
\alias{httk-package}
\alias{httk}
\title{\Sexpr{tools:::Rd_package_title("httk")}}
\description{
\Sexpr{tools:::Rd_package_description("httk")}
}
\seealso{
\href{https://www.epa.gov/chemical-research/computational-toxicology-communities-practice-high-throughput-toxicokinetic-httk}{PowerPoint Presentation: High-Throughput Toxicokinetics (HTTK) R package}

\href{https://www.epa.gov/chemical-research/toxicokinetics-overview}{Toxicokinetics at the U.S. Environmental Protection Agency}

Breen et al. (2021): High-throughput  PBTK models for in vitro to in vivo
extrapolation
(\doi{10.1080/17425255.2021.1935867})

Pearce et al. (2017): httk: R Package for High-Throughput Toxicokinetics
(\doi{10.18637/jss.v079.i04})

Armitage et al. (2014): Application of mass balance models and the chemical
activity concept to facilitate the use of in vitro toxicity data for risk 
assessment
(\doi{10.1021/es501955g})

Honda et al. (2019): Using the concordance of in vitro and in vivo data to
evaluate extrapolation  assumptions
(\doi{10.1371/journal.pone.0217564})

Kapraun et al. (2022): Evaluation of a rapid, generic human gestational dose 
model
(\doi{10.1016/j.reprotox.2022.09.004})

Linakis et al. (2020): Development and evaluation of a high throughput 
inhalation model for organic chemicals
(\doi{10.1038/s41370-020-0238-y})

Pearce et al. (2017): Evaluation and calibration of high-throughput
predictions of chemical distribution to tissues
(\doi{10.1007/s10928-017-9548-7})

Ring et al. (2017): Identifying populations sensitive to environmental
chemicals by simulating toxicokinetic variability
(\doi{10.1016/j.envint.2017.06.004})

Sipes et al. (2017): An Intuitive Approach for Predicting Potential Human
Health Risk with the Tox21 10k Library
(\doi{10.1021/acs.est.7b00650})

Wambaugh et al. (2015): Toxicokinetic Triage for Environmental Chemicals
(\doi{10.1093/toxsci/kfv118})

Wambaugh et al. (2018): Evaluating In Vitro-In Vivo Extrapolation of
Toxicokinetics}
(\doi{10.1093/toxsci/kfy020})

Wambaugh et al. (2019): Assessing Toxicokinetic Uncertainty and Variability
in Risk Prioritization
(\doi{10.1093/toxsci/kfz205})

Wetmore et al. (2015): Incorporating High-Throughput Exposure Predictions with Dosimetry-Adjusted
In Vitro Bioactivity to Inform Chemical Toxicity Testing
(\doi{10.1093/toxsci/kfv171)

\href{https://www.epa.gov/chemical-research/rapid-chemical-exposure-and-dose-research}{The U.S. EPA ExpoCast (Exposure Forecasting) Project}
}
\author{
John Wambaugh, Robert Pearce, Caroline Ring, Gregory Honda, Nisha
Sipes, Jimena Davis, Barbara Wetmore, Woodrow Setzer, Mark Sfeir
}
\keyword{package}
