% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hw_H}
\alias{hw_H}
\title{Optimal bandwidth for kernel density estimate of joint weight/height residual
variability}
\format{
A list with 10 elements, named for unique combinations of NHANES
  gender categories ("Male" and "Female") and NHANES race/ethnicity
  categories ("Mexican American", "Non-Hispanic White", "Non-Hispanic Black",
  "Other", and "Other Hispanic"). Each list element is a 2x2 matrix giving
  the variance-covariance matrix that is the optimal 2-D KDE bandwidth for
  estimating the joint distribution of residual variability in (natural-log)
  weight/height residuals.
}
\source{
The output of \code{\link[ks]{Hpi}} called on the matrix of weight
  and height residuals (in \link{kde_centers}, columns \code{logwresid} and
  \code{loghresid}, with NAs removed)
}
\usage{
hw_H
}
\description{
In virtual-individuals mode, HTTK-Pop generates height and weight for
simulated individuals, by first using the gender- and race/ethnicity-specific
fitted smooothing splines for height and weight to predict values based on
age. Then, HTTK-Pop adds residual variability to the spline-predicted values.
Residual variability is estimated using a kernel-density estimate (KDE),
which is determined in the following way. First, the fitted splines are used
to predict (natural-log) weight and height values for each NHANES respondent.
Then, residuals are calculated as the difference between the natural-log
NHANES measured value and the spline-predicted value. Then, a 2-D Gaussian
kernel distribution is placed centered on each pair of weight/height
residuals. (To account for sample weights, centers are effectively repeated a
number of times corresponding to their relative weight.) The 2-D Gaussian
kernel's bandwidth (variance-covariance matrix) is determinedusing the
plug-in estimator from \code{\link[ks]{Hpi}}. The sum of all the kernel
distributions approximates the distribution of residuals.
}
\details{
To sample from the resulting KDE, the following algorithm is used. First,
draw a random sample with replacement from the "kernel centers", i.e. the
rows of the original matrix of residuals. (Use the NHANES sample weights to
determine the probability of choosing each center.) Then, for each sampled
center (pair of weight/height residuals), draw a random value from a 2-D
normal distribution whose mean is the sampled pair of weight/height
residuals, and whose variance-covariance matrix is the optimal KDE bandwidth.
}
\keyword{datasets}
