% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_mass_functions.R
\name{kidney_mass_children}
\alias{kidney_mass_children}
\title{Predict kidney mass for children}
\usage{
kidney_mass_children(weight, height, gender)
}
\arguments{
\item{weight}{Vector of weights in kg.}

\item{height}{Vector of heights in cm.}

\item{gender}{Vector of genders (either 'Male' or 'Female').}
}
\value{
A vector of kidney masses in kg.
}
\description{
For individuals under age 18, predict kidney mass from weight, height, and
gender. using equations from Ogiu et al. 1997
}
\references{
Ogiu, Nobuko, et al. "A statistical analysis of the internal 
organ weights of normal Japanese people." Health physics 72.3 (1997): 368-383.

Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability."
Environment International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{httk-pop}
