% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kramer.R
\name{kramer_eval}
\alias{kramer_eval}
\title{Evaluate the Kramer In Vitro Distribution model}
\usage{
kramer_eval(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  casrn.vector = NA_character_,
  nomconc.vector = 1,
  this.well_number = 384,
  tcdata = NA,
  user_assay_parameters = NA,
  this.serum = NA_real_,
  this.gKow = NA_real_,
  this.Hconst = NA_real_,
  this.csalt = 0.15,
  this.BSA = 44,
  this.v_total = NA_real_,
  this.v_working = NA_real_,
  this.cell_yield = NA_real_,
  this.L_per_mil_cells = 2.772e-06,
  this.sarea = NA_real_,
  this.Tsys = 37,
  this.Tref = 298.15,
  this.temp_k = 298.15,
  this.prot_conc = 0.21,
  this.option.bottom = TRUE,
  this.option.plastic = FALSE,
  restrict.ion.partitioning = FALSE,
  this.option.piechart = FALSE,
  surface.area.switch = TRUE
)
}
\arguments{
\item{tcdata}{A data.table with casrn, nomconc,v_total, v_working. Otherwise supply single values to this.params (e.g., this.sarea,
this.v_total, etc.). Chemical parameters are taken from 
\code{\link{chem.physical_and_invitro.data}}.}

\item{restrict.ion.partitioning}{only allow neutral fraction to partition}

\item{casrn}{description}

\item{well_number}{description}

\item{gkow_n}{Log octanol to water PC (unitless)}

\item{logHenry}{log10 Henry's law constant (atm*m^3/mol)}

\item{nomconc}{Nominal test concentration (uM)}

\item{v_total}{Total volume of well (uL)}

\item{v_working}{Volume of medium per well (uL)}

\item{cell_yield}{Number of cells/well seeded (unitless)}

\item{sarea}{Surface area of plastic exposed to medium (m^2)}

\item{Tsys}{System temperature (Celcius)}

\item{Tref}{Reference temperature (Kelvin)}

\item{prot_conc}{Cell protein concentration (mg protein/million cells)}

\item{serum}{Concentration of serum in media (percent volume/volume)}

\item{BSA}{BSA concentration in serum (g/L)}

\item{option.piechart}{Option to generate a pie chart to visualize chemical partitioning}
}
\value{
\tabular{lll}{
\strong{Input Parameter} \tab \strong{Description} \tab \strong{Units} \cr
concentration_cells \tab Concentration in cells \tab uM \cr 
concentration_medium \tab Concentration in medium \tab uM \cr 
concentration_plastic \tab Concentration in plastic \tab umol/m^2 \cr 
concentration_air \tab Concentration in headspace \tab uM \cr 
}
}
\description{
Evaluate the Kramer model for chemical distribution \emph{in vitro}. Takes input
as data table or vectors of values. Outputs a data table.
}
\references{
\insertref{kramer2010measuring}
}
\author{
Meredith Scherer, adapted from code written by L.S Lautz for A. Punt, N. Kramer
}
