% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_1comp.R
\name{parameterize_1comp}
\alias{parameterize_1comp}
\title{Parameters for a one compartment (empirical) toxicokinetic model}
\usage{
parameterize_1comp(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  species = "Human",
  default.to.human = FALSE,
  adjusted.Funbound.plasma = TRUE,
  adjusted.Clint = TRUE,
  regression = TRUE,
  restrictive.clearance = TRUE,
  well.stirred.correction = TRUE,
  suppress.messages = FALSE,
  clint.pvalue.threshold = 0.05,
  minimum.Funbound.plasma = 1e-04
)
}
\arguments{
\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- the 
chemical must be identified by either CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  the 
chemical must be identified by either CAS, name, or DTXISD}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
-- the chemical must be identified by either CAS, name, or DTXSIDs}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{default.to.human}{Substitutes missing rat values with human values if
true.}

\item{adjusted.Funbound.plasma}{Uses 
\href{https://doi.org/10.1007/s10928-017-9548-7}{Pearce et al. (2017)} 
lipid binding adjustment
for Funbound.plasma (which impacts volume of distribution) when set to TRUE 
(Default).}

\item{adjusted.Clint}{Uses 
\href{https://doi.org/10.1124/dmd.108.020834}{Kilford et al. (2008)} 
hepatocyte incubation
binding adjustment for Clint when set to TRUE (Default).}

\item{regression}{Whether or not to use the regressions in calculating
partition coefficients in volume of distribution calculation.}

\item{restrictive.clearance}{In calculating elimination rate and hepatic
bioavailability, protein binding is not taken into account (set to 1) in
liver clearance if FALSE.}

\item{well.stirred.correction}{Uses correction in calculation of hepatic
clearance for well-stirred model if TRUE.  This assumes clearance relative
to amount unbound in whole blood instead of plasma, but converted to use
with plasma concentration.}

\item{suppress.messages}{Whether or not to suppress messages.}

\item{clint.pvalue.threshold}{Hepatic clearance for chemicals where the in
vitro clearance assay result has a p-value greater than the threshold are
set to zero.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}
}
\value{
\item{Vdist}{Volume of distribution, units of L/kg BW.}
\item{plasma.vol}{Volume of the plasma, L/kg BW.}
\item{Fgutabs}{Fraction of the oral dose absorbed, that is, the fraction of the
dose that enters the gutlumen.} 
\item{Fhep.assay.correction}{The fraction of chemical unbound in hepatocyte 
assay using the method of Kilford et al. (2008)} 
\item{kelim}{Elimination rate, units of 1/h.} 
\item{hematocrit}{Percent volume of red blood cells in the blood.}
\item{kgutabs}{Rate chemical is absorbed, 1/h.}
\item{million.cells.per.gliver}{Millions cells per gram of liver tissue.}
\item{MW}{Molecular Weight, g/mol.} 
\item{Rblood2plasma}{The ratio of the concentration of the chemical in the 
blood to the concentration in the plasma. Not used in calculations but 
included for the conversion of plasma outputs.} 
\item{hepatic.bioavailability}{Fraction of dose remaining after
first pass clearance, calculated from the corrected well-stirred model.}
\item{BW}{Body Weight, kg.} 
\item{pKa_Donor}{Ionization equilibria (if any) for hydrogen donation (acids).} 
\item{pKa_Accept}{Ionization equilibria (if any) for hydrogen acceptance (bases).} 
\item{Pow}{Octanol:Water equilibrium partition coefficient (hydrophobicity)}
\item{MA}{Membrane affinity -- physopholipid bi-layer partition coefficient}
}
\description{
This function initializes the parameters needed in the function solve_1comp.
The toxiokinetic model is of the form of an
empirical, single compartment in which all tissues are well mixed. 
The route of exposure can be oral or intravenous. 
For oral exposures a hepatic extraction factor (first-pass 
metabolism) is estimated using
chemical-specific \emph{in vitro}-measured intrinsic hepatic 
clearance and fraction
unbound in plasma.
The rate of oral absorption used
is 2.2 L/h, the median rate observed across 44 chemicals by 
\href{https://doi.org/10.1093/toxsci/kfy020}{Wambaugh et al. (2018)}.
There is a single, unspecified route of elimination (clearance). 
Clearance of the chemical from the blood is estimated from 
\emph{in vitro}-measured chemical-specific parameters as described by
\href{https://doi.org/10.1080/17425255.2021.1935867}{Breen et al. (2021)}.  
Volume of distribution is estimated from \emph{in vitro}-measured 
chemical-specific plasma protein binding and physico-chemical parameters
using a modified 
\href{https://doi.org/10.1016/j.tiv.2007.09.010}{Schmitt (2008)} method
to predict tissue-specific particition coefficients 
(\href{https://doi.org/10.1007/s10928-017-9548-7}{Pearce et al. (2017)}) and 
then lumping the tissues into a single compartment weighted by tissue volume:
}
\details{
\if{latex}{
\eqn{V_{d,steady-state} = \Sigma_{i\in tissues}K_{i}V_{i} + V_{plasma}}
}
\if{html}{
V_d,steady-state = Sum over all tissues (K_i * V_i) + V_plasma
}

where K_i is the tissue:unbound plasma concentration partition coefficient
for tissue i.
}
\examples{

 parameters <- parameterize_1comp(chem.name='Bisphenol-A',species='Rat')
 parameters <- parameterize_1comp(chem.cas='80-05-7',
                                  restrictive.clearance=FALSE,
                                  species='rabbit',
                                  default.to.human=TRUE)
 out <- solve_1comp(parameters=parameters)

}
\references{
Pearce, Robert G., et al. "Httk: R package for high-throughput 
toxicokinetics." Journal of statistical software 79.4 (2017): 1.

Schmitt, Walter. "General approach for the calculation of tissue 
to plasma partition coefficients." Toxicology in vitro 22.2 (2008): 457-467.

Pearce, Robert G., et al. "Evaluation and calibration of high-throughput 
predictions of chemical distribution to tissues." Journal of pharmacokinetics 
and pharmacodynamics 44.6 (2017): 549-565.

Kilford, P. J., Gertz, M., Houston, J. B. and Galetin, A.
(2008). Hepatocellular binding of drugs: correction for unbound fraction in
hepatocyte incubations using microsomal binding or drug lipophilicity data.
Drug Metabolism and Disposition 36(7), 1194-7, 10.1124/dmd.108.020834.
}
\seealso{
\code{\link{solve_1comp}}

\code{\link{calc_analytic_css_1comp}}

\code{\link{calc_vdist}}

\code{\link{parameterize_steadystate}}

\code{\link{apply_clint_adjustment}}

\code{\link{tissue.data}}

\code{\link{physiology.data}}
}
\author{
John Wambaugh and Robert Pearce
}
\keyword{1compartment}
\keyword{Parameter}
