% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_armitage.R
\name{parameterize_armitage}
\alias{parameterize_armitage}
\title{Parameterize Armitage In Vitro Distribution Model}
\usage{
parameterize_armitage(
  tcdata = NA,
  casrn.vector = NA_character_,
  nomconc.vector = 1,
  this.well_number = 384,
  this.logKow = NA_real_,
  this.cell_yield = NA_real_,
  this.sarea = NA_real_,
  this.prot_conc = 0.21,
  this.option.bottom = TRUE,
  this.FBSf = NA_real_,
  this.Tsys = 37,
  this.Tref = 298.15,
  this.option.kbsa2 = FALSE,
  this.option.swat2 = FALSE,
  this.pseudooct = 0.01,
  this.memblip = 0.04,
  this.nlom = 0.2,
  this.P_nlom = 0.035,
  this.P_dom = 0.05,
  this.P_cells = 1,
  this.csalt = 0.15,
  this.celldensity = 1,
  this.cellmass = 3,
  this.f_oc = 1,
  this.conc_ser_alb = 24,
  this.conc_ser_lip = 1.9,
  this.Vdom = 0,
  this.option.plastic = FALSE,
  restrict.ion.partitioning = FALSE,
  surface.area.switch = TRUE
)
}
\value{
\tabular{lll}{
\strong{Param} \tab \strong{Description} \tab \strong{Units} \cr
gkaw \tab The air to water PC \tab unitless ratio \cr  
logHenry \tab The log10 Henry's law constant \tab atm*m^3/mol \cr  
gswat \tab The log10 water solubility at 25C (logWSol) \tab log10 mg/L \cr
}
}
\description{
Parameterize Armitage In Vitro Distribution Model
}
\author{
Meredith Scherer
}
