% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_armitage.R
\name{parameterize_armitage}
\alias{parameterize_armitage}
\title{Parameterize Armitage In Vitro Distribution Model}
\usage{
parameterize_armitage(
  tcdata = NA,
  casrn.vector = NA_character_,
  nomconc.vector = 1,
  this.well_number = 384,
  this.cell_yield = NA_real_,
  this.sarea = NA_real_
)
}
\arguments{
\item{tcdata}{A data.table with casrn, nomconc, MP, gkow, gkaw, gswat, sarea,
v_total, v_working. Otherwise supply single values to this.params (e.g., this.sarea,
this.v_total, etc.). Chemical parameters are taken from 
\code{\link{chem.physical_and_invitro.data}}.}

\item{casrn.vector}{A deprecated argument specifying a single or vector of 
Chemical Abstracts Service Registry 
Number(s) (CAS-RN) of desired chemical(s).}

\item{nomconc.vector}{For vector or single value, micromolar (uM = mol/L) nominal 
concentration (e.g. AC50 value)}

\item{this.well_number}{For single value, plate format default is 384, used
if is.na(tcdata)==TRUE. This value chooses default surface area settings for
\code{\link{armitage_estimate_sarea}} based on the number of wells per plate.}

\item{this.cell_yield}{For single value, optionally supply cell_yield,
otherwise estimated based on well number}

\item{this.sarea}{Surface area per well (m^2)}
}
\value{
\tabular{lll}{
\strong{Param} \tab \strong{Description} \tab \strong{Units} \cr
gkaw \tab The air to water PC \tab unitless ratio \cr  
logHenry \tab The log10 Henry's law constant \tab atm*m^3/mol \cr  
gswat \tab The log10 water solubility at 25C (logWSol) \tab log10 mg/L \cr
}
}
\description{
Parameterize Armitage In Vitro Distribution Model
}
\author{
Meredith Scherer
}
