% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameterize_dermal_PBTK.R
\name{parameterize_dermal_pbtk}
\alias{parameterize_dermal_pbtk}
\title{Parameterize_dermal_PBTK}
\usage{
parameterize_dermal_pbtk(
  chem.cas = NULL,
  chem.name = NULL,
  dtxsid = NULL,
  model.type = "dermal",
  method.permeability = "Ficks-Law",
  species = "Human",
  default.to.human = F,
  tissuelist = list(liver = c("liver"), kidney = c("kidney"), lung = c("lung"), gut =
    c("gut"), skin = "skin"),
  force.human.clint.fup = F,
  clint.pvalue.threshold = 0.05,
  adjusted.Funbound.plasma = T,
  regression = T,
  suppress.messages = F,
  minimum.Funbound.plasma = 1e-04,
  skin_depth = 0.3,
  skin.pH = 7,
  vmax.km = F,
  height = 175,
  ...
)
}
\arguments{
\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{https://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{model.type}{Choice of dermal model, either the default "dermal" for the 
model with 2 sub compartments (stratum corneum and viable epidermis) for skin,
or "dermal_1subcomp" for the model with 1 compartment for the skin.}

\item{method.permeability}{For "dermal_1subcomp" model, method of calculating 
the permeability coefficient, Kp, either "Potts-Guy" or "Ficks-Law". Default
is "Ficks-Law" (Sawyer et al., 2016 and Chen and Han et al., 2015), which uses Fick's
law of diffusion to calculate Kp. For "dermal" model, this parameter is ignored.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{default.to.human}{Substitutes missing animal values with human values
if true (hepatic intrinsic clearance or fraction of unbound plasma).}

\item{tissuelist}{Specifies compartment names and tissues groupings.
Remaining tissues in tissue.data are lumped in the rest of the body.
However, solve_dermal_pbtk only works with the default parameters.}

\item{force.human.clint.fup}{Forces use of human values for hepatic
intrinsic clearance and fraction of unbound plasma if true.}

\item{clint.pvalue.threshold}{Hepatic clearances with clearance assays
having p-values greater than the threshold are set to zero.}

\item{adjusted.Funbound.plasma}{Returns adjusted Funbound.plasma when set to
TRUE along with parition coefficients calculated with this value.}

\item{regression}{Whether or not to use the regressions in calculating
partition coefficients.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

\item{skin_depth}{skin_depth of skin, cm, used in calculating Kp.}

\item{skin.pH}{pH of dermis/skin, used in calculating Kp and Kskin2media.}

\item{vmax.km}{Whether or not to use Michaelis-Menten kinetics, returning
Vmax and Km in parameters instead of Clmetabolismc and
million.cells.per.gliver. THIS INPUT IS CURRENTLY NOT USED.}

\item{height}{Height in cm, used in calculating totalSA.}
}
\value{
\item{BW}{Body Weight, kg.} \item{Clmetabolismc}{Hepatic Clearance, L/h/kg
BW.} \item{Fgutabs}{Fraction of the oral dose absorbed, i.e. the fraction of
the dose that enters the gutlumen.} \item{Funbound.plasma}{Fraction of
plasma that is not bound.} \item{Fhep.assay.correction}{The fraction of
chemical unbound in hepatocyte assay using the method of Kilford et al.
(2008)} \item{hematocrit}{Percent volume of red blood cells in the blood.}
\item{Kgut2pu}{Ratio of concentration of chemical in gut tissue to unbound
concentration in plasma.} \item{kgutabs}{Rate that chemical enters the gut
from gutlumen, 1/h.} \item{Kkidney2pu}{Ratio of concentration of chemical in
kidney tissue to unbound concentration in plasma.} \item{Kliver2pu}{Ratio of
concentration of chemical in liver tissue to unbound concentration in
plasma.} \item{Klung2pu}{Ratio of concentration of chemical in lung tissue
to unbound concentration in plasma.} \item{Krbc2pu}{Ratio of concentration
of chemical in red blood cells to unbound concentration in plasma.}
\item{Krest2pu}{Ratio of concentration of chemical in rest of body tissue to
unbound concentration in plasma.} \item{million.cells.per.gliver}{Millions
cells per gram of liver tissue.} \item{MW}{Molecular Weight, g/mol.}
\item{Qcardiacc}{Cardiac Output, L/h/kg BW^3/4.} \item{Qgfrc}{Glomerular
Filtration Rate, L/h/kg BW^3/4, volume of fluid filtered from kidney and
excreted.} \item{Qgutf}{Fraction of cardiac output flowing to the gut.}
\item{Qkidneyf}{Fraction of cardiac output flowing to the kidneys.}
\item{Qliverf}{Fraction of cardiac output flowing to the liver.}
\item{Qlungf}{Fraction of cardiac output flowing to the lung.}
\item{Qskinf}{Fraction of cardiac output flowing to the skin, or to the viable
epidermis layer of the skin when model.type="dermal".}
\item{Rblood2plasma}{The ratio of the concentration of the chemical in the
blood to the concentration in the plasma from available_rblood2plasma.}
\item{Vartc}{Volume of the arteries per kg body weight, L/kg BW.}
\item{Vgutc}{Volume of the gut per kg body weight, L/kg BW.}
\item{Vkidneyc}{Volume of the kidneys per kg body weight, L/kg BW.}
\item{Vliverc}{Volume of the liver per kg body weight, L/kg BW.}
\item{Vlungc}{Volume of the lungs per kg body weight, L/kg BW.}
\item{Vmucc}{Volume of the ? per kg body weight, L/kg BW. (Not used in model?)}
\item{Vrestc}{ Volume of the rest of the body per kg body weight, L/kg BW.}
\item{Vvenc}{Volume of the veins per kg body weight, L/kg BW.}
\item{Vskinc}{Volume of the skin per kg body weight, L/kg BW.}
\item{Vstratum_corneumc}{Volume of the stratum corneum layer of the skin per 
kg body weight, L/kg BW. This parameter does not appear when 
model.type="dermal_1subcomp".}
\item{Vviable_epidermisc}{Volume of the viable epidermis layer of the skin per 
kg body weight, L/kg BW. This parameter does not appear when 
model.type="dermal_1subcomp".}
\item{Kp}{Permeability of the skin, cm/h. When model.type="dermal_1subcomp", 
this parameter changes depending on method.permeability. When model.type="dermal",
this parameter is replaced by Pmedia2sc and Psc2ve.} \item{Pmedia2sc}{Permeability
of the stratum corneum (SC), cm/h. This parameter does not appear when 
model.type="dermal_1subcomp".} \item{Psc2ve}{Permeability of the viable epidermis (VE), cm/h.
This parameter does not appear when model.type="dermal_1subcomp".} \item{Kskin2media}{Partition coefficient
between exposed skin and media. This parameter only appears when model.type="dermal_1subcomp"
and is replaced by Ksc2media when model.type="dermal".} \item{Ksc2media}{Partition 
coefficient between SC and media. This parameter does not appear when 
model.type="dermal_1subcomp".} \item{Kve2media}{Partition coefficient between VE and
SC. This parameter does not appear when model.type="dermal_1subcomp".} \item{totalSA}{Total body surface area,
cm^2.} \item{Vmedia}{Volume of media, L.} \item{skin_depth}{Skin skin_depth, cm.}
\item{Fdermabs}{Fraction of media concentration available for absorption.}
\item{Fskin_exposed}{Fraction of skin exposed.} \item{Vmax}{units/hr, with
units corresponding to dosing concentration. CURRENTLY NOT IN USE.} \item{Km}{units same as dosing
concentration. CURRENTLY NOT IN USE.}
}
\description{
This function initializes the parameters needed in the functions solve_pbtk,
calc_css, and others using the multiple compartment model.
}
\examples{


 

}
\references{
Kilford, P. J., Gertz, M., Houston, J. B. and Galetin, A.
(2008). Hepatocellular binding of drugs: correction for unbound fraction in
hepatocyte incubations using microsomal binding or drug lipophilicity data.
Drug Metabolism and Disposition 36(7), 1194-7, 10.1124/dmd.108.020834.
}
\author{
John Wambaugh and Robert Pearce
}
\keyword{Parameter}
