% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_IVD.R
\name{parameterize_IVD}
\alias{parameterize_IVD}
\title{Parameterize In Vitro Distribution Models}
\usage{
parameterize_IVD(tcdata = NA, casrn.vector = NA_character_, this.pH = 7)
}
\arguments{
\item{tcdata}{A data.table with casrn, nomconc, MP, gkow, gkaw, gswat, sarea,
v_total, v_working. Otherwise supply single values to this.params (e.g., this.sarea,
this.v_total, etc.). Chemical parameters are taken from 
\code{\link{chem.physical_and_invitro.data}}.}

\item{casrn.vector}{For vector or single value, CAS number}

\item{this.pH}{pH of media}
}
\value{
\tabular{lll}{
\strong{Param} \tab \strong{Description} \tab \strong{Units} \cr
casrn \tab Chemical Abstracts Service Registry Number \tab character \cr
logHenry \tab The log10 Henry's law constant \tab atm*m^3/mol \cr      
gswat \tab The log10 water solubility at 25C (logWSol) \tab log10 mol/L \cr
MP_C \tab The chemical compound's melting point \tab degrees C \cr           
MW \tab The chemical compound's molecular weight \tab g/mol \cr          
gkow_n \tab The log10 octanol to water (PC) (logP)\tab log10 unitless ratio \cr  
pKa_Donor \tab Chemical dissociation equilibrium constant(s); pKa(ie pKa_Donor) = -log10(Ka) \tab unitless \cr  
pKa_Accept \tab Chemical association equilibrium constant(s); pKb(ie pKa_Accept) = 14 - pKa  \tab unitless \cr 
pH \tab pH where ionization is evaluated (typically assay medium) \tab unitless \cr 
gkaw_n \tab The air to water PC (neutral) \tab unitless ratio \cr 
}
}
\description{
This function collects physicochemical properties from chemicals input by the user for use with armitage.R and kramer.R.
}
\examples{

library(httk)

output <- parameterize_IVD(casrn.vector = c("15687-27-1"))
print(output)

}
\author{
Meredith Scherer
}
