% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_IVD.R
\name{parameterize_IVD}
\alias{parameterize_IVD}
\title{Parameterize In Vitro Distribution Models}
\usage{
parameterize_IVD(
  tcdata = NA,
  casrn.vector = NA_character_,
  nomconc.vector = NA_real_,
  this.well_number = 384,
  this.sarea = NA_real_,
  this.cell_yield = NA_real_,
  this.v_working = NA_real_,
  this.v_total = NA_real_,
  this.pH = 7
)
}
\arguments{
\item{tcdata}{A data.table with casrn, nomconc, MP, gkow, gkaw, gswat, sarea,
v_total, v_working. Otherwise supply single values to this.params (e.g., this.sarea,
this.v_total, etc.). Chemical parameters are taken from 
\code{\link{chem.physical_and_invitro.data}}.}

\item{casrn.vector}{For vector or single value, CAS number}

\item{nomconc.vector}{For vector or single value, micromolar (uM = mol/L) nominal 
concentration (e.g. AC50 value)}

\item{this.well_number}{For single value, plate format default is 384, used
if is.na(tcdata)==TRUE. This value chooses default surface area settings for
\code{\link{armitage_estimate_sarea}} based on the number of plates per well.}

\item{this.sarea}{Surface area of plastic exposed to medium (m^2)}

\item{this.cell_yield}{Number of cells per well}

\item{this.v_working}{Working volume per well (uL)}

\item{this.v_total}{Total volume per well (uL)}
}
\value{
\tabular{lll}{
\strong{Param} \tab \strong{Description} \tab \strong{Units} \cr
casrn \tab Chemical Abstracts Service Registry Number \tab character \cr
well_number \tab Number of wells in plate (used to set default surface area) \tab unitless \cr   
sarea \tab Surface area of plastic exposed to medium \tab m^2 \cr         
v_total \tab Total volume of well \tab uL \cr       
v_working \tab Filled volume of well \tab uL \cr     
cell_yield \tab Number of cells \tab cells \cr
gkow \tab The log10 octanol to water (PC) (logP)\tab log10 unitless ratio \cr     
logHenry \tab The log10 Henry's law constant \tab atm*m^3/mol \cr      
gswat \tab The log10 water solubility at 25C (logWSol) \tab log10 mol/L \cr
MP \tab The chemical compound's melting point \tab degrees C \cr           
MW \tab The chemical compound's molecular weight \tab g/mol \cr            
}
}
\description{
This function collects information on chemicals and well plate geometry input by the user.
}
\examples{

library(httk)

output <- parameterize_IVD(casrn.vector = c("15687-27-1"), this.well_number = 384)
print(output)

}
\author{
Meredith Scherer
}
