% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_kramer.R
\name{parameterize_kramer}
\alias{parameterize_kramer}
\title{Parameterize Kramer IVD Model}
\usage{
parameterize_kramer(
  tcdata = NA,
  casrn.vector = NA_character_,
  nomconc.vector = 1,
  this.well_number = 384,
  this.BSA = 44,
  this.logKow = NA_real_,
  this.logHenry = NA_real_,
  this.serum = NA_real_,
  this.v_total = NA_real_,
  this.v_working = NA_real_,
  this.cell_yield = NA_real_,
  this.sarea = NA_real_,
  this.temp_k = 298.15,
  this.prot_conc = 0.21,
  this.option.bottom = TRUE,
  this.option.plastic = FALSE,
  restrict.ion.partitioning = FALSE,
  surface.area.switch = TRUE
)
}
\arguments{
\item{tcdata}{A data table with well_number corresponding to plate format,
optionally include v_working, sarea, option.bottom, and option.plastic}

\item{casrn.vector}{A deprecated argument specifying a single or vector of 
Chemical Abstracts Service Registry 
Number(s) (CAS-RN) of desired chemical(s).}

\item{nomconc.vector}{For vector or single value, micromolar (uM = mol/L) nominal 
concentration (e.g. AC50 value)}

\item{this.well_number}{For single value, plate format default is 384, used
if is.na(tcdata)==TRUE}

\item{this.BSA}{bovine serum albumin concentration (g/L)}

\item{this.v_total}{For single value, optionally supply total volume,
otherwise estimated based on well number (uL)}

\item{this.v_working}{For single value, optionally supply working volume,
otherwise estimated based on well number (uL)}

\item{this.cell_yield}{For single value, optionally supply cell_yield,
otherwise estimated based on well number}
}
\value{
A data table composed of any input data.table \emph{tcdata}
with only the following columns either created or altered by this function:
\tabular{ccc}{
\strong{Column Name} \tab \strong{Description} \tab \strong{Units} \cr
well_number \tab number of wells on plate \tab \cr
sarea \tab surface area \tab m^2 \cr
cell_yield \tab number of cells \tab cells \cr
v_working_m3 \tab working (filled) volume of each well \tab uL \cr
v_total_m3 \tab total volume of each well \tab uL \cr
}
}
\description{
This function takes inputs from kramer_eval() and calls parameterize_IVD(). Converts units and sets up variables for kramer_eval().
}
\author{
Meredith Scherer
}
