% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pearce2017regression}
\alias{pearce2017regression}
\alias{Pearce2017Regression}
\title{Pearce et al. 2017 data}
\format{
data.frame
}
\source{
Pearce et al. 2017 Regression Models
}
\usage{
pearce2017regression
}
\description{
This table includes the adjusted and unadjusted regression parameter
estimates for the chemical-specifc plasma
protein unbound fraction (fup) in 12 different tissue types.
}
\details{
Predictions were made with regression models, 
as reported in Pearce et al. (2017).
}
\references{
Pearce, Robert G., et al. "Evaluation and calibration of 
high-throughput predictions of chemical distribution to tissues."
Journal of pharmacokinetics and pharmacodynamics 44.6 (2017): 549-565.
}
\seealso{
\code{\link{predict_partioning_schmitt}}
}
\author{
Robert G. Pearce
}
\keyword{data}
