% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dosing.R
\name{scale_dosing}
\alias{scale_dosing}
\title{Scale mg/kg body weight doses according to body weight and units}
\usage{
scale_dosing(dosing, parameters, route)
}
\arguments{
\item{dosing}{List of dosing metrics used in simulation, which includes
the namesake entries of a model's associated dosing.params. In the case
of most httk models, these should include "initial.dose", "doses.per.day", 
"daily.dose", and "dosing.matrix". The "dosing.matrix" is used for more
precise dose regimen specification, and is a matrix consisting of two
columns or rows named "time" and "dose" containing the time and amount,
in mg/kg BW, of each dose.}

\item{parameters}{Chemical parameters, either from parameterization 
function, or otherwise specified. Must contain a BW entry.}

\item{route}{String specification of route of exposure for simulation:
"oral", "iv", "inhalation", ...}
}
\value{
A list of numeric values for doses converted to output.units, potentially
(depending on argument dosing) including:
\item{initial.dose}{The first dose given}
\item{dosing.matrix}{A 2xN matrix where the first column is dose time and
the second is dose amount for N doses}
\item{daily.dose}{The total cumulative daily dose}
}
\description{
This function scales doses (in mg/kg) by body weight. It handles single 
doses, matrices of doses, or daily repeated doses at varying intervals. 
Gut absorption is also factored in through the parameter Fgutabs,
and scaling is currently avoided in the inhalation exposure case with a 
scale factor of 1
}
\author{
John Wambaugh, Matt Linakis, and Mark Sfeir
}
\keyword{Dynamic}
