% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dosing.R
\name{scale_dosing}
\alias{scale_dosing}
\title{Scale mg/kg body weight doses according to body weight and units}
\usage{
scale_dosing(
  dosing,
  parameters,
  route,
  input.units = NULL,
  output.units = "uM",
  vol = NULL
)
}
\arguments{
\item{dosing}{List of dosing metrics used in simulation, which must include
the general entries with names "initial.dose", "doses.per.day", 
"daily.dose", and "dosing.matrix". The "dosing.matrix" is used for more
precise dose regimen specification, and is a matrix consisting of two
columns or rows named "time" and "dose" containing the time and amount,
in mg/kg BW, of each dose. The minimal usage case involves all entries but
"initial.dose" set to NULL in value.}

\item{parameters}{Chemical parameters from parameterize_pbtk function,
overrides chem.name and chem.cas.}

\item{route}{String specification of route of exposure for simulation:
"oral", "iv", "inhalation", ...}

\item{input.units}{Units of the dose values being scaled. (Default is NULL.) 
Currently supported units "mg/L", "ug/L","ug/mL", "uM", "umol/L", "ug/dL",
"ug/g", "nmol/L", "nM", and "ppmw" (supported input.units subject to change).}

\item{output.units}{Desired units (either "mg/L", "mg", "umol", or default
"uM").}

\item{vol}{Volume for the target tissue of interest.
NOTE: Volume should not be in units of per BW, i.e. "kg".}
}
\value{
A list of numeric values for doses converted to output.units, potentially
(depending on argument dosing) including:
\item{initial.dose}{The first dose given}
\item{dosing.matrix}{A 2xN matrix where the first column is dose time and
the second is dose amount for N doses}
\item{daily.dose}{The total cumulative daily dose}
}
\description{
This function transforms the dose (in mg/kg) into the appropriate units. It
handles single doses, matrices of doses, or daily repeated doses at varying
intervals. Gut absorption is also factored in through the parameter Fgutabs,
and scaling is currently avoided in the inhalation exposure case with a 
scale factor of 1
}
\author{
John Wambaugh and Sarah E. Davidson
}
\keyword{Dynamic}
