% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipes2017}
\alias{sipes2017}
\alias{Sipes2017}
\title{Sipes et al. 2017 data}
\format{
data.frame
}
\source{
ADMET, Simulations Plus
}
\usage{
sipes2017
}
\description{
This table includes in silico predicted chemical-specifc plasma protein 
unbound fraction (fup) and intrinsic hepatic clearance values for the entire
Tox21 library 
(see \url{https://www.epa.gov/chemical-research/toxicology-testing-21st-century-tox21}). 
Predictions were made with Simulations Plus ADMET predictor,
as reported in Sipes et al. (2017).
}
\references{
Sipes, Nisha S., et al. "An Intuitive Approach for Predicting
Potential Human Health Risk with the Tox21 10k Library." Environmental
Science & Technology 51.18 (2017): 10786-10796.
}
\author{
Nisha Sipes
}
\keyword{data}
