% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_1tri_pbtk.R
\name{solve_1tri_pbtk}
\alias{solve_1tri_pbtk}
\title{Solve_1tri_PBTK}
\usage{
solve_1tri_pbtk(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  times = seq(0, 13 * 7, 1),
  parameters = NULL,
  days = NULL,
  species = "human",
  tsteps = 4,
  dose = NULL,
  dosing.matrix = NULL,
  daily.dose = NULL,
  doses.per.day = NULL,
  initial.values = NULL,
  plots = FALSE,
  suppress.messages = FALSE,
  iv.dose = FALSE,
  input.units = "mg/kg",
  output.units = NULL,
  method = "lsoda",
  rtol = 1e-08,
  atol = 1e-12,
  default.to.human = FALSE,
  recalc.blood2plasma = FALSE,
  recalc.clearance = FALSE,
  adjusted.Funbound.plasma = TRUE,
  regression = TRUE,
  restrictive.clearance = TRUE,
  minimum.Funbound.plasma = 1e-04,
  monitor.vars = NULL,
  ...
)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{times}{Optional time sequence in days. Dosing sequence begins at the
beginning of times. Default is from 0th week of pregnancy to 13th due to 
model representation.}

\item{parameters}{Chemical parameters from parameterize_fetal_pbtk function,
overrides chem.name and chem.cas.}

\item{days}{Length of the simulation.}

\item{species}{Included for compatibility with other functions, but the model
will not run for non-human species (default "Human").}

\item{tsteps}{The number time steps per hour. Default of 4.}

\item{dose}{Amount of a single, initial oral dose in mg/kg BW.}

\item{dosing.matrix}{A matrix of either one column (or row) with a set of
dosing times or with two columns (or rows) correspondingly named "dose" and
"time" containing the time and amount, in mg/kg BW, of each dose.}

\item{daily.dose}{Total daily dose, mg/kg BW.}

\item{doses.per.day}{Number of doses per day.}

\item{initial.values}{Vector containing the initial concentrations or
amounts of the chemical in specified tissues with units corresponding to
compartment.units.  Defaults are zero.}

\item{plots}{Plots all outputs if true.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{iv.dose}{Simulates a single i.v. dose if true.}

\item{input.units}{Input units of interest assigned to dosing, defaults to
mg/kg BW}

\item{output.units}{A named vector of output units expected for the model
results. Default, NULL, returns model results in units specified in the
'modelinfo' file. See table below for details.}

\item{method}{Method used by integrator (deSolve).}

\item{rtol}{Argument passed to integrator (deSolve).}

\item{atol}{Argument passed to integrator (deSolve).}

\item{default.to.human}{Substitutes missing animal values with human values
if true (hepatic intrinsic clearance or fraction of unbound plasma).}

\item{recalc.blood2plasma}{Recalculates the ratio of the amount of chemical
in the blood to plasma using the input parameters, calculated with
hematocrit, Funbound.plasma, and Krbc2pu.}

\item{recalc.clearance}{Recalculates the the hepatic clearance
(Clmetabolism) with new million.cells.per.gliver parameter.}

\item{adjusted.Funbound.plasma}{Uses adjusted Funbound.plasma when set to
TRUE along with partition coefficients calculated with this value.}

\item{regression}{Whether or not to use the regressions in calculating
partition coefficients.}

\item{restrictive.clearance}{Protein binding not taken into account (set to
1) in liver clearance if FALSE.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

\item{monitor.vars}{Which variables to track by default}

\item{...}{Additional arguments passed to the integrator.}
}
\value{
A matrix of class deSolve with a column for time(in days), each
compartment, the area under the curve, and plasma concentration and a row
for each time point.
}
\description{
This function solves for the amounts or concentrations in uM of a chemical
in different tissues of a pregnant woman as functions of time based on
the dose and dosing frequency.
}
\details{
The model begins by default at non-pregnancy (0th week) and ends at the 13th 
week of pregnancy, thereby simulating the 1st trimester. This is meant to 
augment the fetal_pbtk model (Kapraun et al. 2022) which is limited to the 
13th to 40th week window. 

Note that the model parameters have units of hours while the model output is
in days. Dose is in mg, not scaled for body weight.

Default NULL value for doses.per.day solves for a single dose.

The maternal compartments used in this model are the gut lumen, gut, liver, 
venous blood, arterial blood, lung, adipose tissue, kidney, thyroid, brain, 
and rest of body. The "conceptus" compartment models an early developing fetus 
along with the products of conception (i.e. placenta, amniotic fluid) through 
which chemical exchange can occur with the maternal blood. 

The extra compartments include the amounts or concentrations metabolized by
the liver and excreted by the kidneys through the tubules.

AUC is the area under the curve of the plasma concentration.

This gestational model is only parameterized for humans.
}
\examples{

Note: These examples should be worked out after testing the functions. 
out = solve_fetal_pbtk(chem.name = 'bisphenol a', daily.dose = 1,
doses.per.day = 3)

# With adjustement to fraction unbound plasma for fetus:
fetal_parms_fup_adjusted <- 
  parameterize_fetal_pbtk(chem.name = 'perfluorooctane sulfonic acid')
head(solve_fetal_pbtk(parameters = fetal_parms_fup_adjusted))

# Without adjustement to fraction unbound plasma for fetus:
fetal_parms_fup_unadjusted <-  
  parameterize_fetal_pbtk(chem.name = 'perfluorooctane sulfonic acid',
                                fetal_fup_adjustment = FALSE)
head(solve_fetal_pbtk(parameters = fetal_parms_fup_unadjusted))


}
\seealso{
\code{\link{solve_model}}

\code{\link{parameterize_1tri_pbtk}}
}
\author{
John Wambaugh, Mark Sfeir, and Dustin Kapraun
}
\keyword{Solve}
