% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_dermal_pbtk.R
\name{solve_dermal_pbtk}
\alias{solve_dermal_pbtk}
\title{Solve_dermal_PBTK}
\usage{
solve_dermal_pbtk(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  times = NULL,
  parameters = NULL,
  days = 10,
  species = "Human",
  tsteps = 4,
  dose = NULL,
  dosing.matrix = NULL,
  forcings = NULL,
  exp.start.time = 0,
  exp.conc = 1,
  exp.duration = 12,
  period = 24,
  daily.dose = NULL,
  doses.per.day = NULL,
  initial.values = NULL,
  plots = F,
  suppress.messages = F,
  input.units = "mg/kg",
  output.units = NULL,
  method = "lsoda",
  rtol = 1e-08,
  atol = 1e-12,
  default.to.human = F,
  recalc.blood2plasma = F,
  recalc.clearance = F,
  adjusted.Funbound.plasma = T,
  regression = T,
  restrictive.clearance = T,
  minimum.Funbound.plasma = 1e-04,
  monitor.vars = NULL,
  skin_depth = 0.3,
  skin.pH = 7,
  vmax.km = F,
  ...
)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs}

\item{times}{Optional time sequence for specified number of days.  Dosing
sequence begins at the beginning of times.}

\item{parameters}{Chemical parameters from parameterize_pbtk function,
overrides chem.name and chem.cas.}

\item{days}{Length of the simulation.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{tsteps}{The number time steps per hour.}

\item{dosing.matrix}{Matrix consisting of three columns named
"concentration", "Vmedia", and "time" containing the dosing times, days,
with the applied concentration, units/L, and the volume of the applied
media, L.}

\item{doses.per.day}{Number of doses per day.}

\item{initial.values}{Vector containing the initial concentrations or
amounts of the chemical in specified tissues with units corresponding to
output.units.  Defaults are zero.}

\item{plots}{Plots all outputs if true.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{method}{Method used by integrator (deSolve).}

\item{rtol}{Argument passed to integrator (deSolve).}

\item{atol}{Argument passed to integrator (deSolve).}

\item{default.to.human}{Substitutes missing animal values with human values
if true (hepatic intrinsic clearance or fraction of unbound plasma).}

\item{recalc.blood2plasma}{Recalculates the ratio of the amount of chemical
in the blood to plasma using the input parameters, calculated with
hematocrit, Funbound.plasma, and Krbc2pu.}

\item{recalc.clearance}{Recalculates the the hepatic clearance
(Clmetabolism) with new million.cells.per.gliver parameter.}

\item{adjusted.Funbound.plasma}{Uses adjusted Funbound.plasma when set to
TRUE along with partition coefficients calculated with this value.}

\item{regression}{Whether or not to use the regressions in calculating
partition coefficients.}

\item{restrictive.clearance}{Protein binding not taken into account (set to
1) in liver clearance if FALSE.}

\item{skin_depth}{skin_depth of skin, cm, used in calculating Kp.}

\item{skin.pH}{pH of dermis/skin, used in calculating Kp and Kskin2media.}

\item{vmax.km}{Whether or not to use Michaelis-Menten kinetics}

\item{...}{Additional arguments passed to the integrator.}

\item{concentration}{Concentration of media at dosing times, units/L.}

\item{use.amounts}{Return outputs as amounts instead of concentrations.}
}
\value{
A matrix of class deSolve with a column for time(in days), each
compartment, the area under the curve, and plasma concentration and a row
for each time point.
}
\description{
This function solves for the amounts or concentrations of a chemical in
different tissues as functions of time based on the dose and dosing
frequency.
}
\details{
Model units are the same as media concentration, units/L or units when
use.amounts=T.

New doses replace rather than add to previous doses. A concentration of 0 in
dosing.matrix switches off the dosing/diffusion between the media and
exposed skin.

Note that the model parameters have units of hours while the model output is
in days.

Default NULL value for doses.per.day solves for a single dose.

The compartments used in this model are the gutlumen, gut, liver, kidneys,
veins, arteries, lungs, skin, exposed skin, media, and the rest of the body.

The extra compartments include the amounts or concentrations metabolized by
the liver and excreted by the kidneys through the tubules.

AUC is the area under the curve of the plasma concentration.

When species is specified as rabbit, dog, or mouse, the function uses the
appropriate physiological data(volumes and flows) but substitues human
fraction unbound, partition coefficients, and intrinsic hepatic clearance.
}
\examples{

Vmedia <- c(.1,.2,.3)
time <- c(0,.5,3)
concentration <- c(2,0,3)
dosing.matrix <- cbind(time,concentration,Vmedia)
out <- solve_dermal_pbtk(chem.name='bisphenola',dosing.matrix=dosing.matrix,plots=T)

parameters <- parameterize_dermal_pbtk(chem.name='bisphenola',skin_depth=1)
parameters$Fskinexposed <- 0.25
parameters$Kp <- 2
parameters$Fdermabs <- 0.5
parameters$Vmedia <- 1
out <- solve_dermal_pbtk(parameters=parameters,concentration=100,plots=T)

}
\author{
John Wambaugh and Robert Pearce
}
\keyword{Solve}
