% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_dermal_pbtk.R
\name{solve_dermal_pbtk}
\alias{solve_dermal_pbtk}
\title{Solve_dermal_PBTK}
\usage{
solve_dermal_pbtk(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  model.type = "dermal_1subcomp",
  method.permeability = "Chen-Lian",
  Kvehicle2water = NULL,
  times = NULL,
  parameters = NULL,
  days = 10,
  tsteps = 4,
  plots = FALSE,
  monitor.vars = NULL,
  suppress.messages = F,
  species = "Human",
  method = "lsoda",
  rtol = 1e-08,
  atol = 1e-12,
  recalc.blood2plasma = FALSE,
  recalc.clearance = FALSE,
  adjusted.Funbound.plasma = TRUE,
  minimum.Funbound.plasma = 1e-04,
  parameterize.arg.list = list(default.to.human = FALSE, clint.pvalue.threshold = 0.05,
    restrictive.clearance = TRUE, regression = TRUE),
  route = NULL,
  Vvehicle = NULL,
  initial.dose = NULL,
  input.units = NULL,
  dose.duration = NULL,
  dose.duration.units = NULL,
  dosing.dermal = NULL,
  dosing.matrix = NULL,
  washoff = FALSE,
  InfiniteDose = FALSE,
  ...
)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})
the chemical must be identified by either CAS, name, or DTXSIDs.}

\item{model.type}{Choice of dermal model, either the default "dermal_1subcomp" for
the model with 1 compartment for the skin; or (not usable yet) "dermal" for the 
model with 2 sub compartments (a top and bottom layer) for skin which defaults 
to the top layer being the stratum corneum and the bottom layer being the combined
viable epidermis and dermis.}

\item{method.permeability}{For "dermal_1subcomp" model, method of calculating 
the permeability coefficient, P, either "Potts-Guy" or "Chen-Lian". Default
is "Chen-Lian" (Sawyer et al., 2016 and Chen et al., 2015), which uses Fick's
law of diffusion to calculate P. For "dermal" model, this parameter is ignored.}

\item{Kvehicle2water}{Partition coefficient for the vehicle (sometimes called the 
vehicle) carrying the chemical to water. Default is "water", which assumes the vehicle is water.
Other optional inputs are "octanol" and "olive oil".}

\item{times}{Optional time sequence for specified number of days.  Dosing
sequence begins at the beginning of times.}

\item{parameters}{Chemical parameters from parameterize_dermal_pbtk function,
overrides chem.name and chem.cas.}

\item{days}{Length of the simulation.}

\item{tsteps}{The number time steps per hour.}

\item{plots}{Plots all outputs if true.}

\item{monitor.vars}{Which variables are returned as a function of time. 
Default values of NULL looks up variables specified in modelinfo_MODEL.R}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{method}{Method used by integrator (deSolve).}

\item{rtol}{Argument passed to integrator (deSolve).}

\item{atol}{Argument passed to integrator (deSolve).}

\item{recalc.blood2plasma}{Recalculates the ratio of the amount of chemical
in the blood to plasma using the input parameters, calculated with
hematocrit, Funbound.plasma, and Krbc2pu.}

\item{recalc.clearance}{Recalculates the the hepatic clearance
(Clmetabolism) with new million.cells.per.gliver parameter.}

\item{adjusted.Funbound.plasma}{Uses adjusted Funbound.plasma when set to
TRUE along with partition coefficients calculated with this value.}

\item{parameterize.arg.list}{Additional parameterized passed to the model 
parameterization function, "parameterize_dermal_pbtk". The inputs "model.type",
"method.permeability", and "Kvehicle2water" are not passed through this.}

\item{Vvehicle}{Volume of vehicle applied to skin in L, defaults to 0.01 L. If InfiniteDose
is TRUE, this parameter is ignored and set = 1.}

\item{initial.dose}{Concentration}

\item{dosing.dermal}{Matrix consisting of three columns named
"Cvehicle", "Vvehicle", and "time" containing the dosing times, days,
with the applied amount in the vehicle, and the volume of the applied
vehicle, L. Note that the units of Cvehicle are controlled by input.units. ***If
InfiniteDose=TRUE, the Vvehicle column of dosing.dermal is ignored.}

\item{washoff}{This parameter only matters if dosing.dermal is being used. If 
TRUE, any chemical left on the skin is assumed to be replaced be new dose. If
FALSE (default), any chemical left on the skin is added to the new dose.}

\item{...}{Additional arguments passed to the integrator.}

\item{BW}{Body weight, kg.}

\item{InitialDose}{If TRUE, we assume infinite dosing (i.e., a constant unchanging concentration
of chemical in the vehicle is considered) and Cvehicle is a constant. If
FALSE (default), dosing is finite and Cvehicle changes over time.}
}
\value{
A matrix of class deSolve with a column for time (in days), each
compartment, the area under the curve, and plasma concentration and a row
for each time point.
}
\description{
This function solves for the amounts or concentrations of a chemical in
different tissues as functions of time based on the dose and dosing
frequency.
}
\details{
Model units are the same as vehicle concentration, units/L or units when
use.amounts=T.

New doses replace rather than add to previous doses. A concentration of 0 in
dosing.matrix switches off the dosing/diffusion between the vehicle and
exposed skin.

Note that the model parameters have units of hours while the model output is
in days.

Default NULL value for doses.per.day solves for a single dose.

The compartments used in this model are the gutlumen, gut, liver, kidneys,
veins, arteries, lungs, unexposed stratum corneum and viable epidermis, exposed 
stratum corneum and viable epidermis, vehicle, and the rest of the body. When model.type 
= "dermal_1subcomp", the stratum corneum and viable epidermis are replaced with 
one compartment: skin.

The extra compartments include the amounts or concentrations metabolized by
the liver and excreted by the kidneys through the tubules.

AUC is the area under the curve of the plasma concentration.

When species is specified as rabbit, dog, or mouse, the function uses the
appropriate physiological data(volumes and flows) but substitutes human
fraction unbound, partition coefficients, and intrinsic hepatic clearance.
}
\examples{

Vvehicle <- c(.1,.2,.3)
time <- c(0,.5,3)
Cvehicle <- c(2,0,3)
dosing.dermal <- cbind(time,Cvehicle,Vvehicle)
out <- solve_dermal_pbtk(chem.name='bisphenola',dosing.dermal=dosing.dermal,plots=T)

parameters <- parameterize_dermal_pbtk(chem.name='bisphenola',skin_depth=1)
parameters$Fskin_exposed <- 0.25
parameters$Vvehicle <- 1
out <- solve_dermal_pbtk(parameters=parameters,plots=T)

}
\author{
Annabel Meade, John Wambaugh, and Robert Pearce
}
\keyword{Solve}
