% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_full_pregnancy.R
\name{solve_full_pregnancy}
\alias{solve_full_pregnancy}
\title{Solve_full_pregnancy}
\usage{
solve_full_pregnancy(
  dtxsid,
  track.vars = NULL,
  plt = FALSE,
  return.units = "amt",
  time.course = seq(0, 40 * 7, 1),
  ...
)
}
\arguments{
\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})}

\item{track.vars}{which variables to return in solution output dataframe}

\item{plt}{plots all outputs, if TRUE}

\item{return.units}{if plt = TRUE, plots outputs in desired units of interest.}

\item{time.course}{Time sequence in days. Default is from 0th week of pregnancy to 
40th, incremented by day.}

\item{...}{additional arguments passed to solve_fetal_pbtk and solve_1tri_pbtk
(this is where you input daily.dose and doses.per.day)}
}
\value{
A dataframe with columns for time (in days), each compartment, the 
area under the curve (for plasma vs time), and plasma, and a row for each time 
point.
}
\description{
This function solves for the amounts (in umol) or concentrations (in uM) of a 
chemical in different tissues of a maternal-fetal system over the full course 
of human pregnancy given a dose and dosing frequency.
}
\details{
The simulation starts at the 0th week and ends at 40 weeks of pregnancy (term), 
covering all trimesters of human pregnancy. This is accomplished by 
stitching together the 1tri and fetal PBTK models with appropriate initial 
conditions, as described in Truong et al. (TBD).
}
\seealso{
\code{\link{solve_1tri_pbtk}}

\code{\link{solve_fetal_pbtk}}

\code{\link{parameterize_1tri_pbtk}}

\code{\link{parameterize_fetal_pbtk}}
}
\author{
Kimberly Truong
}
\keyword{Solve}
