% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_full_pregnancy.R
\name{solve_full_pregnancy}
\alias{solve_full_pregnancy}
\title{Solve_full_pregnancy}
\usage{
solve_full_pregnancy(
  chem.name = NULL,
  chem.cas = NULL,
  dtxsid = NULL,
  time.course = seq(0, 40 * 7, 1),
  track.vars = NULL,
  plt = FALSE,
  return.units = "amt",
  ...
)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or DTXSID
must be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or DTXSID must be specified.}

\item{dtxsid}{EPA's DSSTox Structure ID (\url{http://comptox.epa.gov/dashboard})}

\item{time.course}{Time sequence in days. Default is from 0th week of pregnancy to 
40th, incremented by day.}

\item{track.vars}{which variables to return in solution output dataframe}

\item{plt}{plots all outputs, if TRUE}

\item{return.units}{if plt = TRUE, plots outputs in desired units of interest
(either 'amt' for chemical amounts or 'conc' for chemical concentration).}

\item{...}{additional arguments passed to solve_fetal_pbtk and solve_1tri_pbtk
(this is where you input daily.dose and doses.per.day)}
}
\value{
A dataframe with columns for time (in days), each compartment, the 
area under the curve (for plasma vs time), and plasma, and a row for each time 
point.
}
\description{
This function solves for the amounts (in umol) or concentrations (in uM) of a 
chemical in different tissues of a maternal-fetal system over the full course 
of human pregnancy given a dose and dosing frequency.
}
\details{
The simulation starts at the 0th week and ends at 40 weeks of pregnancy (term), 
covering all trimesters of human pregnancy. This is accomplished by 
stitching together the 1tri and fetal PBTK models with appropriate initial 
conditions, as described in Truong et al. (TBD).
}
\examples{

# dosing schedule of 1 mg/kg BW/day for 40 weeks
# return solution by hour
out <- solve_full_pregnancy(dtxsid = "DTXSID4034609", # Fipronil 
                               daily.dose = 1, 
                               doses.per.day = 1,
                               time.course = seq(0, 40*7, 1/24))
                   
                               
# return solution in chemical amounts for fetal compartments + placenta
maternal_compts <- c('gutlumen', 'gut', 'liver', 'kidney', 'lung', 'ven', 'art', 
'adipose','thyroid', 'rest')

fetal_compts <- c(maternal_compts[! maternal_compts \%in\% c('adipose', 'gutlumen') ], 
"brain")

amt.out <- solve_full_pregnancy(dtxsid = "DTXSID4034609", # Fipronil 
                               daily.dose = 1, 
                               doses.per.day = 1,
                               time.course = seq(0, 40*7, 1), 
                               track.vars = c(paste0("Af", fetal_compts), "Aplacenta"))

# return solution in concentrations for fetal compartments + placenta 
conc.out <- solve_full_pregnancy(dtxsid = "DTXSID4034609", # Fipronil 
                                daily.dose = 1, 
                                doses.per.day = 1,
                                time.course = seq(0, 40*7, 1), 
                                track.vars = c(paste0("Cf", fetal_compts), "Cplacenta"))

# plot solution in units of amounts
# in this case, time.course affects both deSolve output and plot 
gplot.out <- solve_full_pregnancy(dtxsid = "DTXSID5022308", # Genistein
                                 daily.dose = 1, 
                                 doses.per.day = 1,
                                 time.course = seq(89,92, 1/24), 
                                 plt = T, return.units = "amt")

}
\seealso{
\code{\link{solve_1tri_pbtk}}

\code{\link{solve_fetal_pbtk}}

\code{\link{parameterize_1tri_pbtk}}

\code{\link{parameterize_fetal_pbtk}}
}
\author{
Kimberly Truong
}
\keyword{Solve}
