% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue.data}
\alias{tissue.data}
\title{Tissue composition and species-specific physiology parameters}
\format{
A data.frame containing 406 rows and 5 columns.
\tabular{ll}{
\strong{Column} \tab \strong{Description} \cr
Tissue \tab The tissue being described \cr                      
Species \tab The species being described \cr                      
Reference \tab The reference for the value reported \cr                      
variable \tab The aspect of the tissue being characterized \cr                      
value \tab The value for the variable for the given tissue and species \cr                      
}
}
\usage{
tissue.data
}
\description{
This data set contains values from Schmitt (2008) and Ruark et al. (2014)
describing the composition of specific tissues and from Birnbaum et al.
(1994) describing volumes of and blood flows to those tissues, allowing
parameterization of toxicokinetic models for human, mouse, rat, dog, or
rabbit. Tissue volumes were calculated by converting the fractional mass of
each tissue with its density (both from ICRP), lumping the remaining tissues
into the rest-of-body, excluding the mass of the gastrointestinal contents.
}
\details{
Many of the parameters were compiled initially in Table 2 of Schmitt (2009).
The full list of tissue variables described is:
\tabular{lll}{
\strong{Variable} \tab \strong{Description} \tab \strong{Units} \cr
Fcell \tab Cellular fraction of total tissue volume \tab fraction \cr
Fint \tab Interstitial fraction of total tissue volume \tab fraction \cr
FWc \tab Fraction of cell volume that is water \tab fraction \cr
FLc \tab Fraction of cell volume that is lipid \tab fraction \cr
FPc \tab Fraction of cell volume that is protein \tab fraction \cr
Fn_Lc \tab Fraction of cellular lipid tht is neutral lipid \tab fraction \cr
Fn_PLc \tab Fraction of cellular lipid tht is neutral phospholipid \tab fraction \cr
Fa_PLc \tab Fraction of cellular lipid tht is acidic phospholipid \tab fraction \cr
pH \tab Negative logarithm of H+ ion concentration \tab unitless\cr
Density \tab Tissue density \tab g/cm^3 \cr
Vol \tab Tissue volume \tab L/kg \cr
Flow \tab Blood flow to tissue \tab mL/min/kg^(3/4) \cr
}

New tissues can be added to this table to generate
their partition coefficients.
}
\examples{
# We can add thyroid to the tissue data by making a row containing
# its data, subtracting the volumes and flows from the rest-of-body, 
# and binding the row to tissue.data. Here we assume it contains the same 
# partition coefficient data as the spleen and a tenth of the volume and  
# blood flow:
new.tissue <- subset(tissue.data,Tissue == "spleen")
new.tissue[, "Tissue"] <- "thyroid"
new.tissue[new.tissue$variable \%in\% c("Vol (L/kg)",
"Flow (mL/min/kg^(3/4))"),"value"] <- new.tissue[new.tissue$variable
\%in\% c("Vol (L/kg)","Flow (mL/min/kg^(3/4))"),"value"] / 10
tissue.data[tissue.data$Tissue == "rest", "value"] <-
tissue.data[tissue.data$Tissue == "rest", "value"] -
new.tissue[new.tissue$variable \%in\% c("Vol (L/kg)",
"Flow (mL/min/kg^(3/4))"),"value"]
tissue.data <- rbind(tissue.data, new.tissue)

# We can add a new species (for example, wolverines) by adding new information
# to the physiology.data and tissue.data tables. It can be convenient to start by
# by replicating the data from another species and adjusting as appropriate:

# Copy physiology data from rabbit:
new.species <- physiology.data[,"Rabbit"]
names(new.species) <- physiology.data[,"Parameter"]
rabbit.BW <- new.species["Average BW"] 
new.species["Average BW"] <- 31.2 # Rausch and Pearson (1972) https://doi.org/10.2307/3799057
new.species["Average Body Temperature"] <- 38.5 # Thiel et al. (2019) https://doi.org/10.1186/s12983-019-0319-8

# Add new physiology data column to physiology.data table"
physiology.data <- cbind(physiology.data, new.species)
colnames(physiology.data)[length(colnames(physiology.data))] <- "Wolverine"

# Copy tissue data from rabbit:
new.tissue.data <- subset(tissue.data,Species=="Rabbit")
new.tissue.data$Species <- "Wolverine"

# Add new tissue data rows to tissue.data table:
tissue.data <- rbind(tissue.data, new.tissue.data)

# Species is now available for calculations:
calc_mc_css(chem.cas="80-05-7",
            species="wolverine",
            parameterize.arg.list=list(default.to.human=TRUE),
            suppress.messages=TRUE,
            samples = 100)

}
\references{
\insertRef{birnbaum1994physiological}{httk}

\insertRef{ruark2014predicting}{httk}

\insertRef{schmitt2008general}{httk}

\insertRef{snyder1974report}{httk}

\insertRef{wambaugh2015toxicokinetic}{httk}
}
\seealso{
\code{\link{predict_partitioning_schmitt}}
}
\author{
John Wambaugh, Robert Pearce, and Nisha Sipes
}
\keyword{data}
