% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue.data}
\alias{tissue.data}
\title{Tissue composition and species-specific physiology parameters}
\format{
A data.frame containing 13 rows and 20 columns.
}
\usage{
tissue.data
}
\description{
This data set contains values from Schmitt (2008) and Ruark et al. (2014)
describing the composition of specific tissues and from Birnbaum et al.
(1994) describing volumes of and blood flows to those tissues, allowing
parameterization of toxicokinetic models for human, mouse, rat, dog, or
rabbit. Tissue volumes were calculated by converting the fractional mass of
each tissue with its density (both from ICRP), lumping the remaining tissues
into the rest-of-body, excluding the mass of the gastrointestinal contents
}
\details{
New tissues can be added to this table to generate
their partition coefficients.

The tissue data needed for calculating partition coefficients include: 
cellular and water fractions of 
total volume, lipid and protein  fractions of cellular volume, lipid 
fractions of the total lipid volume, the pH of each tissue,
and the fractional volume of protein in plasma.
}
\examples{
# We can add thyroid to the tissue data by making a row containing
# its data, subtracting the volumes and flows from the rest-of-body, 
# and binding the row to tissue.data. Here we assume it contains the same 
# partition coefficient data as the spleen and a tenth of the volume and  
# blood flow:
new.tissue <- subset(tissue.data,Tissue == "spleen")
new.tissue[, "Tissue"] <- "thyroid"
new.tissue[new.tissue$variable \%in\% c("Vol (L/kg)",
"Flow (mL/min/kg^(3/4))"),"value"] <- new.tissue[new.tissue$variable
\%in\% c("Vol (L/kg)","Flow (mL/min/kg^(3/4))"),"value"] / 10
tissue.data[tissue.data$Tissue == "rest", "value"] <-
tissue.data[tissue.data$Tissue == "rest", "value"] -
new.tissue[new.tissue$variable \%in\% c("Vol (L/kg)",
"Flow (mL/min/kg^(3/4))"),"value"]
tissue.data <- rbind(tissue.data, new.tissue)

# We can add a new species (for example, wolverines) by adding new information
# to the physiology.data and tissue.data tables. It can be convenient to start by
# by replicating the data from another species and adjusting as appropriate:

# Copy physiology data from rabbit:
new.species <- physiology.data[,"Rabbit"]
names(new.species) <- physiology.data[,"Parameter"]
rabbit.BW <- new.species["Average BW"] 
new.species["Average BW"] <- 31.2 # Rausch and Pearson (1972) https://doi.org/10.2307/3799057
new.species["Average Body Temperature"] <- 38.5 # Thiel et al. (2019) https://doi.org/10.1186/s12983-019-0319-8

# Add new physiology data column to physiology.data table"
physiology.data <- cbind(physiology.data, new.species)
colnames(physiology.data)[length(colnames(physiology.data))] <- "Wolverine"

# Copy tissue data from rabbit:
new.tissue.data <- subset(tissue.data,Species=="Rabbit")
new.tissue.data$Species <- "Wolverine"

# Add new tissue data rows to tissue.data table:
tissue.data <- rbind(tissue.data, new.tissue.data)

# Species is now available for calculations:
calc_mc_css(chem.cas="80-05-7",
            species="wolverine",
            parameterize.arg.list=list(default.to.human=TRUE),
            suppress.messages=TRUE,
            samples = 100)

}
\references{
Birnbaum, L and Brown, R and Bischoff, K and Foran, J and
Blancato, J and Clewell, H and Dedrick, R (1994). Physiological parameter
values for PBPK model. International Life Sciences Institute, Risk Science
Institute, Washington, DC

Ruark, Christopher D., et al. "Predicting passive and active tissue: plasma
partition coefficients: Interindividual and interspecies variability."
Journal of pharmaceutical sciences 103.7 (2014): 2189-2198.

Schmitt, W. (2008). General approach for the calculation of tissue to plasma
partition coefficients. Toxicology in vitro : an international journal
published in association with BIBRA 22(2), 457-67,
10.1016/j.tiv.2007.09.010.

ICRP. Report of the Task Group on Reference Man. ICRP Publication 23 1975

Wambaugh, John F., et al. "Toxicokinetic triage for environmental
chemicals." Toxicological Sciences (2015): 228-237.
}
\seealso{
\code{\link{predict_partitioning_schmitt}}
}
\author{
John Wambaugh, Robert Pearce, and Nisha Sipes
}
\keyword{data}
