% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.seem3}
\alias{wambaugh2019.seem3}
\title{ExpoCast SEEM3 Consensus Exposure Model Predictions for Chemical Intake Rates}
\format{A data frame with 385 rows and 38 variables:}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.seem3
}
\description{
These data are a subset of the Bayesian inferrences reported by Ring et al.
(2019) for a consensus model of twelve exposue predictors. The predictors
were calibrated based upon their ability to predict intake rates inferred
National Health and Nutrition Examination Survey (NHANES). They reflect the
populaton median intake rate (mg/kg body weight/day), with uncertainty.
}
\references{
Ring, Caroline L., et al. "Consensus modeling of median chemical
intake for the US population based on predictions of exposure pathways."
Environmental science & technology 53.2 (2018): 719-732.

Wambaugh et al. (2019) "Assessing Toxicokinetic Uncertainty and
Variability in Risk Prioritization", Toxicological Sciences, 172(2), 235-251.
}
\author{
John Wambaugh
}
\keyword{data}
